Hangman v0.9.1 CLI

Module handles command line interpreter interface to application.

Able to run interactive human games with manually specified secrets and also able to run human and robot games with randomly generated secrets.

Robot games are auto-guessed based on simple strategy heuristics.

Game archival can be captured through logging, e.g. —log option

Usage: --name (player id) --type ("human" or "robot") --random (num random secrets, max 10) [--secret (hangman word(s)) --baseline] [--log --display]

or

Aliase Usage: -n (player id) -t ("human" or "robot") -r (num random secrets, max 10) [-s (hangman word(s)) -bl] [-l -d]

Functions

main(args)

Specs

main([String.t]) :: :ok

Gateway function to fetch and validate parameters. Handles display of help information. Proceeds to run the hangman game from the command line.