Hangman v0.9.1 Chunks

Module to handle word list Chunks for a given length key.
Internally maintains standardized containers for word lists and keeps track of total word counts and number of containers.

Splits big words list into smaller more manageable list Chunks.

The need for chunking arises when we may have arbitrary long word lists/streams, so we chunk the word list to a standard size of 500 words.

A single Chunks abstraction can contain a single chunked word list or multiple chunked word lists.

Chunks provide more manageability especially when we store into the database because the abstraction automatically binaries data leaving a smaller footprint.

Primary functions are new/2, count/1, add/2, and get_words_lazy/1.

Types

binary_chunk :: {binary, integer}

t

Functions

add(chunks, value)

Specs

add(t, binary_chunk) :: t

Takes an existing Chunks and adds the passed in binary chunk tuple.

Heavily used in reduce methods to add a {binary, word_count} to the Chunks accumulator value.

The tuple head is a binaried word list and the tail is the word count

container_size()

Specs

container_size :: pos_integer

count(chunks)

Specs

count(t) :: integer

Returns total word count across all chunk containers

get_words_lazy(chunks)

Specs

get_words_lazy(t) :: Enumerable.t

Returns words in lazy enumerable fashion.

info(chunks)

Specs

info(t) :: Keyword.t

Returns Chunks information

key(chunks)

Specs

key(t) :: pos_integer

Returns Chunks word length key

new(length_key)

Specs

new(pos_integer) :: t

Returns new empty Chunks abstraction

new(length_key, words)

Specs

new(pos_integer, Enumerable.t) :: t

Returns new Chunks abstraction. Does this by splitting and encapsulating words lists from enumerable into standardized chunk containers.
Word lists are binaried for compactness.

size(chunks)

Specs

size(t) :: integer

Return total number of chunk containers