Hangman v0.9.1 Chunks.Stream

Module implements transform routine to take a grouped words stream and return a chunked words stream

Similiar functionality to Dictionary.File transform handlers but more closely associated with Chunks.

Functions

transform(stream, atom1, atom2)

Specs

transform(Enumerable.t, :atom, :atom) :: Enumerable.t

Transforms grouped word stream of the form {length_key, word, group_index}

to chunked word stream of the form {[word1, word2, ..., wordN], word_count}

Returns stream Enumerable.t