Hangman v0.9.1 Counter

Module implements a set of functions for working with letter frequency tallies.

Conventional Hangman strategy deals exclusively with letter frequency data.

Counter is a key value store where a key is a letter string and a value is a positive integer.

Types

key :: String.t

t

value :: pos_integer

Functions

add_letters(counter, word)

Specs

add_letters(t, [] | String.t | Enumerable.t) :: t

Adds letters to Counter without checking for duplicate letters. Returns an updated Counter.

add_unique_letters(counter, word)

Specs

add_unique_letters(t, String.t | [String.t]) :: t

Adds letters to Counter. Ensure letters are unique. Returns an updated Counter

add_words(counter, words)

Specs

add_words(t, words :: Enumerable.t) :: t

Adds words list to Counter. Converts words to char list and then letter counts list. List is then submitted to map for bulk value update. Faster than many tiny map updates

add_words(counter, words_stream, exclusion_set)

Specs

add_words(t, stream :: Enumerable.t, exclusion :: map) :: t

Routines adds word stream to Counter. Splits word stream into manageable lists, relies on add_words/2 to reduce into Counter.

delete(counter)

delete(counter, letters)

Specs

delete(t, none | [] | [String.t]) :: t

Returns an updated Counter after deleting specified keys in letters

empty?(counter)

Specs

empty?(t) :: boolean

Returns true or false whether Counter is empty

equal?(c1, c2)

Specs

equal?(t, t) :: boolean

Returns true if Counters equal

inc_by(counter, key, value \\ 1)

Specs

inc_by(t, key, value) :: t

Increment value for a given key by the given value. Default increment value 1.

items(counter)

Specs

items(t) :: [tuple]

Returns a key-value list of {letter, count} tuples

most_common(counter, n)

Specs

most_common(t, pos_integer) :: [tuple]

Returns list of the most common n codepoint keys and codepoint values

most_common_key(counter, n)

Specs

most_common_key(t, pos_integer) :: list

Returns list of the most common n codepoint keys

new()

Returns new, empty Counter

new(word)

Specs

new(String.t | [tuple] | map) :: t

new(none | []) :: t

Returns new Counter that reflects contents of either String.t, [tuple], map