Hangman v0.9.1 Dictionary

Module defines Dictionary common attributes and types functionality. Serves as a central point to access or update such attributes

Types

kind :: :regular | :big

transform :: :unsorted | :sorted | :grouped | :chunked

Functions

big()

Specs

big :: kind

Returns big type

chunked()

Specs

chunked :: transform

Returns chunked type

chunks_file_delimiter()

Delimiter token used to delimit chunk values in binary chunks file

grouped()

Specs

grouped :: transform

Returns grouped type

paths()

Returns Dictionary.File paths map, arranged by types regular and big

regular()

Specs

regular :: kind

Returns regular type

sorted()

Specs

sorted :: transform

Returns sorted type

unsorted()

Specs

unsorted :: transform

Returns unsorted type