Hangman v0.9.1 Dictionary.File

Provides abstraction for various dictionary files.

The original dictionary file is transformed into intermediate files as cached representations of each transformation layer. Unless the original file changes, transformation every time isn’t necessary since the intermediate files are stored on disk.

Load time is only determined by the last transformed chunk file, which is optimized for ETS load.

Transformation types are unsorted to sorted, sorted to grouped, and grouped to chunked. Each transform handler encapsulates each transform procedure

Functions

make_file_transform(fn_transform)

Specs

make_file_transform((path :: String.t, file :: pid -> String.t)) :: (String.t, String.t -> String.t)

Function builder routine to return the customized transform function

When each returned function runs, its reads in the “untransformed” file, applies the transform lambda, and then writes out to the new path

transform(path, arg, kind)

Specs

transform(path :: String.t, pair :: {Dictionary.transform, Dictionary.transform}, kind :: Dictionary.kind) :: String.t

Takes input file, applies a transform type and returns new file path For example, can be used to first sort a file, then upon second invocation, group that file

transform_handler(path, new_path, atom)

Specs

transform_handler(String.t, String.t, Dictionary.transform) :: String.t

Specific handler implementations for sort, group, and chunk transform