Hangman v0.9.1 Dictionary.File.Stream

Module for accessing input file stream for various original and transformed dictionary file types ranging from unsorted, sorted, grouped, chunked

Types

mode :: :read

t

Functions

delete(fstream)

Specs

delete(t) :: t

Deletes fstream, returns empty file stream

file_handler(fstream, arg)

Specs

file_handler(t, {mode, Dictionary.transform}) :: Enumerable.t

Handles file specific input streams, wrapping underlying file formats. Returns file stream, closes file when finished. Normally accessed through gets_lazy/1.

gets_lazy(fstream)

Specs

gets_lazy(t) :: Enumerable.t

Returns input file stream

new(arg, path)

Specs

new({mode, Dictionary.transform}, String.t) :: t

Returns a new empty file stream