Hangman v0.9.1 Game

Module to create and manage Hangman game abstractions.

Naturally serves as the symbolic core of the game application.

The game embodies the structure of play in this context of letter and word guessing. Therefore it maintains the secret pattern state along with the guess state, delineated by correct and incorrect, letter and word.

Further, a game wouldn't amount to much without a score. If a game is lost the score is automatically 25, as deemed by the threshhold of the particular max incorect guesses. If the number of incorrect letter guesses exceeds the threshhold, the game is lost. The lower the score and hence the fewer rounds to guess the word, the more desirable aim.

The Game handles a single Hangman game or multiple Hangman games and runs each runs sequentially. Therefore only one game is in play at any one time.

Primary game functions are load/3, guess/2, status/1.

Types

code ::
  :game_start |
  :game_keep_guessing |
  :game_won |
  :game_lost |
  :games_over |
  :game_reset

Game status code values

data :: {id, Round.code, code, pattern :: String.t, status :: String.t, summary}

returned Game data

id :: String.t

Game id is String.t (currently using player's name as unique key)

result :: {t, data}

Game result tuple returned to Game.Server

summary ::
  [] |
  [status: :games_over, average_score: integer, games: pos_integer, results: [tuple]]

Game summary values, either [] if we are still playing or [...] if game finished

t

Functions

empty?(game)

Specs

empty?(t) :: boolean

Returns boolean whether Game is empty

equal?(game1, game2)

Specs

equal?(t, t) :: boolean

Returns boolean whether Game states are identical

finished?(game)

Specs

finished?(t) :: boolean

Return boolean whether Game is finished

guess(game, arg)

Specs

guess(t, guess :: Guess.t) :: result

Runs guess against Game secret. Updates Hangman pattern, status, and other game recordkeeping structures.

Guesses follow two types

{:guess_letter, letter} - If correct, returns the :correct_letter data tuple along with game data otherwise, returns the :incorrect_letter data tuple along with game data

{:guess_word, word} - If correct, returns the :correct_word data tuple along with game If incorrect, returns the :incorrect_word data tuple with game data

info(g)

Specs

info(t) :: Keyword.t

Returns Game information

load(id_key, secret, max_wrong)

Specs

load(id, String.t | [String.t], pos_integer) :: t

Loads a new Game state given new secret(s)

mystery_letter()

Specs

mystery_letter :: String.t

status(game)

Specs

status(t) :: {id, code, String.t}

Returns current Game status text