Hangman v0.9.1 Game.Pid.Cache

Module implements GenServer functionality to provide Game.Server pid caching. Ultimately relies on the gproc module to administer registry facilities.

Pid caching prevents a Game.Server process from having to be loaded from scratch on every use

Functions

get_server_pid(player_name, secret)

Specs

get_server_pid(name :: String.t, args :: String.t) :: pid

Checks registry cache for Game.Server pid given unique id, returns cached pid or if not found returns new game pid. Handles race conditions

start_link()

Specs

start_link :: {:ok, pid}

GenServer start link wrapper function