Hangman v0.9.1 Game.Server

Module handles Hangman Game serving to multiple clients.
Each player's active game state is maintained, until the player process exits.

In the event the player aborts abnormally, the player's game state is maintained but removed from active status.

NOTE: The server runs each game one at a time and handles support for multiple games concurrently but is currently not harnessed by the Game.Pid.Cache.Server for initial simplicity purposes and therefore not presently utilized.

Types

id :: String.t

t

Functions

guess(game_pid, key, arg)

Specs

guess(pid, Player.key, Guess.t) :: tuple

Issues guess letter or word request, returns guess result

Internally, runs guess against game secret. Updates Hangman pattern, status, and other game recordkeeping structures.

Guesses follow two types

{:guess_letter, letter} - If correct, returns the :correct_letter data tuple along with game info otherwise, returns the :incorrect_letter data tuple along with game info

{:guess_word, word} - If correct, returns the :correct_word data tuple along with game info If incorrect, returns the :incorrect_word data tuple with game info

initiate_and_length(game_pid, key)

Specs

initiate_and_length(pid, Player.key) :: tuple

Initiates link with client and return secret length Retrieves Game secret length

load(game_pid, id_key, secret, max_wrong \\ 5)

Specs

load(pid, id, String.t | [String.t], pos_integer) :: :ok

Loads new game into server process state. Used primarily by Game.Pid.Cache

start_link(id_key, secret, max_wrong \\ 5)

Specs

start_link(id, String.t | [String.t], pos_integer) :: {:ok, pid}

Start public interface method with secret(s)

status(game_pid, key)

Specs

status(pid, Player.key) :: tuple

Retrieves Game status data

stop(game_pid)

Specs

stop(pid) :: tuple

Issues request to stop GenServer

whereis(id_key)

Specs

whereis(id) :: pid | :atom

Routine returns game server pid from process registry using gproc If not found, returns :undefined