Hangman v0.9.5 Hangman.Action.Human

Implements human action player specific functionality

In Hangman we have two players. One explict - the one guessing, the other implicit, 'the game', 'the user tracking the penalties'.

In this instance, the Action is merely the user making and choosing the guess selections.

The human action is given the choice of the top letter choices to choose from and is able to make an interactive guess.

Types

t :: %Hangman.Action.Human{display: term, round: term, strategy: term, type: term}

Functions

guess(human, guess)

Specs

guess(t, Guess.t) :: tuple | no_return

Routine for :human player type. Performs validation of human guess, executes guess and returns round status

info(human)

Specs

info(t) :: Keyword.t

setup(human)

Specs

setup(t) :: tuple

Sets up round by running a reduction pass. Returns top letter choices to be presented to human