Hangman v0.9.5 Hangman.Action.Robot

Implements robot action player specific functionality

The robot action is reliant on the game strategy to automatically self select the best guess.

Types

t :: %Hangman.Action.Robot{display: term, round: term, strategy: term, type: term}

Functions

guess(robot, data)

Specs

guess(t, Guess.t) :: tuple

Routine for :robot player type. Performs auto-generated guess, returns round status

info(robot)

Specs

info(t) :: Keyword.t

setup(robot)

Specs

setup(t) :: tuple

Sets up round by running a reduction pass. Informs Strategy of reduction pass result to be later used in the guess method.