Hangman v0.9.5 Hangman.CLI.Handler

Module drives Player.Controller, while setting up the proper Game server and Event consumer states beforehand.

Simply stated it politely nudges the player to proceed to the next course of action or make the next guess. The handler also collects input from the user as necessary and displays data back to the user.

When the game is finished it politely ends the game playing.

This module is the goto destination after all the arguments have been collected in Hangman.CLI

Functions

run(name, type, secrets, log, display, guess_timeout \\ 10)

Function run connects all the player specific components together and runs the player game