Elixir Hangman v0.1.1 Hangman.Chunks

Module to handle Hangman word list chunks for a given length

Chunks big words list into smaller more manageable list chunks

Encapsulates raw stream consisting of binary chunks

Types

binary_chunk :: {binary, integer}

t :: %Hangman.Chunks{chunk_count: term, key: term, raw_stream: term, word_count: term}

Functions

add(chunks, v)

Specs

add(t, binary_chunk) :: t

Takes an existing chunk stream and a tuple value The tuple head is a binary chunk and the tail is the number of words

get_count(chunks, atom)

Specs

get_count(t, :atom) :: integer

get_count(t, :atom) :: integer

Performs constant time lookup of number of chunks in stream

get_key(chunks)

Specs

get_key(t) :: pos_integer

Returns word length key associated with Chunks abstraction

get_words_lazy(chunks)

Specs

get_words_lazy(t) :: Enumerable.t

Applies binary unpack function to raw stream then flattens result Returns stream of words (un-binaried)

info(c)

Specs

info(t) :: Keyword.t

Returns chunks information

new(length_key)

Specs

new(pos_integer) :: t

Returns new empty chunks abstraction

new(length_key, words)

Specs

new(pos_integer, Enumerable.t) :: t

Returns new chunks abstraction, populated with stream encapsulation Stream data is split up into chunks and word lists are binaried for compactness for storage purposes