Hangman v0.9.1 Hangman.Counter

Hangman.Counter A set of functions for working with letter frequency counters Counters are key value stores where keys are letter strings and values are positive integers

Types

key :: String.t

t

value :: pos_integer

Functions

add_letters(counter, word)

Specs

add_letters(t, Enumerable.t) :: t

add_letters(t, String.t) :: t

add_letters(t, []) :: t

add_unique_letters(counter, word)

Specs

add_unique_letters(t, String.t) :: t

add_words(counter, words)

Specs

add_words(t, [String.t]) :: t

add_words(counter, words_stream, exclusion_set)

Specs

add_words(t, Enumerable.t, map) :: t

delete(counter)

Specs

delete(t) :: t

delete(counter, letters)

Specs

delete(t, [String.t]) :: t

delete(t, []) :: t

empty?(counter)

Specs

empty?(t) :: boolean

equal?(c1, c2)

Specs

equal?(t, t) :: boolean

inc_by(counter, key, value \\ 1)

Specs

inc_by(t, key, value) :: t

items(counter)

Specs

items(t) :: [tuple]

most_common(counter, n)

Specs

most_common(t, pos_integer) :: [tuple]

most_common_key(counter, n)

Specs

most_common_key(t, pos_integer) :: list

new()

Specs

new :: t

new(word)

Specs

new(map) :: t

new([tuple]) :: t

new(String.t) :: t

new([]) :: t