Hangman v0.9.6 Hangman.Dictionary

Module provides central point to perform dictionary functions for clients for types, paths, random words, and lookup functionality.

Types

kind :: :regular | :big

Functions

directory_path(opts)

Specs

directory_path(Keyword.t) :: String.t | no_return

Returns dictionary dir path

ingestion_enabled(opts)

Specs

ingestion_enabled(Keyword.t) :: boolean | no_return

Returns whether ingestion is enabled

key_range()

lookup(atom, count)

Specs

lookup(:random | :tally | :words, pos_integer) ::
  [String.t] |
  Counter.t |
  Words.t

Dictionary lookup routines

The allowed modes: * :random - extracts specified count of random hangman words. * :tally - retrieve letter tally associated with word length key * :words - retrieve the word data lists associated with the word length key

max_random_words()

random(count)

Specs

random(String.t) :: [String.t] | no_return

Returns random word secrets given count

startup_params(opts)

stop()

Handles dictionary server termination