Hangman v0.9.5 Hangman.Dictionary.Cache behaviour

Module implements a GenServer process providing access to a dictionary word cache. Handles lookup routines to access words, tallys, and random words.

Serves as a wrapper around dictinary specific implementation

Functions

lookup(pid, atom, count)

Specs

lookup(pid, atom, pos_integer) ::
  [String.t] |
  Counter.t |
  Words.t |
  no_return

Cache lookup routines

The allowed modes: * :random - extracts count number of random hangman words. * :tally - retrieve letter tally associated with word length key * :words - retrieve the word data lists associated with the word length key

setup?()

Specs

setup? :: atom | no_return

Check whether ets is setup

start_link(args)

Specs

start_link(Keyword.t) :: {:ok, pid}

GenServer start link wrapper function

stop(pid)

Specs

stop(none | pid) :: {}

Routine to stop server normally

Callbacks

init(arg0)

Specs

init(Keyword.t) :: tuple

GenServer callback to initalize server process

Kicks off ingestion process to load dictionary words