Elixir Hangman v0.1.1 Hangman.Dictionary.Cache.Server behaviour

Functions

lookup(atom, count)

Specs

lookup(atom, pos_integer) :: {}

lookup(atom, pos_integer) :: {}

lookup(atom, pos_integer) :: {}

Lookup routine to retrieve the word data chunk associated with the word length key Uses global server name to retrieve the server pid

lookup(pid, atom, count)

Specs

lookup(pid, atom, pos_integer) :: {}

lookup(pid, atom, pos_integer) :: {}

lookup(pid, atom, pos_integer) :: {}

Lookup routine to retrieve the word data chunk

setup(args)

Specs

setup(Keyword.t) :: :ok | no_return

Sets up ets table with dictionary words arranged in chunks, tallies arranged by word length size, and tags random words. Normalizes the original dictionary file to be loaded into the ets table

start_link(args \\ [normal_dictionary: true])

Specs

start_link(Keyword.t) :: {:ok, pid}

GenServer start_link wrapper function

stop()

stop(pid)

Specs

stop(none) :: {}

stop(pid) :: {}

Routine to stop server normally

Callbacks

handle_call(atom, pid, {})

Specs

handle_call(:atom, pid, {}) :: {}

handle_call({:atom, pos_integer}, {}, {}) :: {}

handle_call({:atom, pos_integer}, {}, {}) :: {}

handle_call({:atom, pos_integer}, {}, {}) :: {}

GenServer callback to stop server normally

init(arg0)

Specs

init(Keyword.t) :: {}

GenServer callback to initalize server process

terminate(reason, {})

Specs

terminate(reason :: term, {}) :: term | no_return

GenServer callback to cleanup server state