Hangman v0.9.6 Hangman.Dictionary.ETS

Module provides access to a dictionary word cache powered by ETS. Provides lookup routines to access words, tallys, and random words. Handles all ETS specific operations while Dictionary serves as a wrapper.

Dictionary.ETS loads the dictionary words into words and stores them into ETS via Dictionary.Ingestion. Upon startup, letter frequency tallies are computed and stored compactly into ETS. Words identified as random are tagged and stored as well.

Functions

dump(path)

Specs

dump(binary) :: :ok | no_return

Dumps ets table into an ets file

get(atom, count)

Specs

get(:random | :counter | :words, pos_integer) ::
  [String.t] |
  Hangman.Counter.t |
  Hangman.Words.t |
  no_return

Get has three modes: :random, :counter, :words

:random - Retrieves all words tagged as random upon startup, then randomly chooses count words from this set, and returns the shuffled words result.

:counter - Retrieves dictionary tally counter given word length key

:words - Retrieves words given word length key Reduces over all the same word keys to aggregate the final words list (using ets bag type)

info()

Specs

info :: atom | no_return

load(path)

Specs

load(binary) :: :dictionary_table | no_return

Loads ets into memory via ets file

new()

Specs

new :: atom

put(atom, arg)

Specs

put(:words | :random | :counter, {pos_integer, [String.t] | Hangman.Counter.t}) ::
  :ok |
  no_return

Put function inserts data into the ETS via three modes: words, random, counter

:words - For each words list, insert into ets

:random - For each list of words and length key, within valid length key sizes, extract @random_words_per_list count words, dedup extracted set and insert into ets

:counter - Inserts verified counter structure into ets

Example key is {:counter, 8} Example {length, counter} is: {8, %Counter{entries: %{"a" => 14490, "b" => 4485, "c" => 7815, "d" => 8046, "e" => 19600, "f" => 2897, "g" => 6009, "h" => 5111, "i" => 15530, "j" => 384, "k" => 2628, "l" => 11026, "m" => 5793, "n" => 12186, "o" => 11462, "p" => 5763, "q" => 422, "r" => 14211, "s" => 16560, "t" => 11870, "u" => 7377, "v" => 2156, "w" => 2313, "x" => 662, "y" => 3395, "z" => 783}}}

setup?()

Specs

setup? :: atom | no_return

Check whether ets is setup