Hangman v0.9.5 Hangman.Dictionary.File.Reader behaviour

Module for accessing input file stream for various original and transformed dictionary file types ranging from original, sorted, grouped, chunked

Types

t

Functions

delete(reader)

Specs

delete(t) :: t

Deletes reader, returns empty file stream

new(type, path)

Specs

new(Hangman.Dictionary.transform, String.t) :: t

Returns a new empty file reader

proceed(reader)

Specs

proceed(t) :: Enumerable.t

Proceed returns input file enumerable

Callbacks

handler(t)

Specs

handler(t) :: Enumerable.t

Handles file specific input streams, wrapping underlying file formats. Returns file enumerable, closes file when finished.