Hangman v0.9.5 Hangman.Dictionary.Ingestion

Module handles the ingestion of hangman dictionary words through the coordination of Ingestion.First.Flow and Ingestion.Cache.Flow

Saves ingestion state in intermediary cache partition files and finally in ets dump file

Module transforms the dictionary file in two steps

a preprocessing step, which ingests the dictionary file and generates ingestion cache partition files

ingestion of cache files into ETS

if both steps are done, we generate an ets table file which can be loaded upon startup the next time through

For the first step, we run flow and store the results in partitioned files. We pass a map of file_pids keyed by word length key, to facilitate writing to the intermediate cache files

After the inital run with the dictionary file, we run flow on the cached output files (the intermediate partition files) saving the initial flow intermediary processing -- windowing etc. Allowing us to generate and load the relevant data (word list chunks, random word generation, tally generation) into ETS concurrently.

Lastly, once both the flows have finished we generate an ets file. On subsequent runs, this bypasses extra flow processing as the ets file is loaded to create the ets

Functions

cleanup(cache_dir, key_file_map)

Specs

cleanup(binary, map) :: :ok

Cleans up open file handles left over from writing to the cached files. Also generates a partition manifest file signifying the initial pass has been completed

delimiter(atom)

process(arg)

Specs

process({:new, binary, binary, map} | {:cache, binary, binary} | {:full, binary, binary, map, binary}) :: :ok

Process method supports two modes: new, cache, and full

New runs the initial ingestion concurrently chunking the original dictionary file into key based partition files which contain the various windowed data

Cache runs a flow against the cached partitioned files and concurrently generates and loads all the relevent information into memory

Full basically invokes new and cache

Full runs the full ingestion process by first running the initial ingestion flow process followed by a state cleanup, then running the ingestion cache flow process

run(args)

Specs

run(Keyword.t) :: :ok

Routine kicks off the ingestion process by setting up the proper state and then running it

If we have pregenerated an ets table file previously use that to load the ets and bypass flow processing

setup(arg)

Specs

setup({:ets, binary} | {:flow, binary, binary, binary}) ::
  {:full, binary, binary, map, binary} |
  {:cache, binary, binary} |
  :ok

Setup has two modes: a) :ets b) :flow

:ets Checks if there is an ets cache file that has been pre-generated. If so, we can avoid running the flow computations because the table is ready to load from the file.

:flow Loads environment to run initial ingestion pass through the dictionary file. If this is the first time through, we set up intermediary ingestion cache files If not, we ensure the partition manifest file is present indicating the initial run is complete and execute the cache ingestion run loading chunks and generating random words into ETS as well as generating tally data