Hangman v0.9.5 Hangman.Dictionary.Transformer

The original Dictionary file is transformed into intermediary representations. Given an original dictionary file f1, this file may be transformed a few times until it is suitable to be loaded into ETS. E.g. f1 -> f2 -> f3 -> f4.

This sequence of transforms (t1, t2, t3, t4) is done initially and does not need to be repeated unless the original file changes.

Dictionary word load time is only determined by the last transformed file f4, which is optimized for ETS load.

Types

t

Functions

new(kind)

Specs

new(Hangman.Dictionary.kind) :: t

Store the transform run functions by order First we sort, then group, then chunk

run(state)

Specs

run(t) :: String.t

Runs all the file transform functions in a succinct way

Delay the actual function invocation to this method We are making the assumption that the enumerable will be processed from keys 1 to 3

NOTE: Need stronger guarentee check