Elixir Hangman v0.1.1 Hangman.Game

Module defines game abstraction which can handle a single hangman game or multiple hangman games

Runs each game or set of games sequentially Therefore only one game is in play at any one time

Primary functions are load, guess, status

Types

id :: String.t

result :: {t, tuple}

t :: %Hangman.Game{correct_letters: term, current: term, id: term, incorrect_letters: term, incorrect_words: term, max_wrong: term, pattern: term, patterns: term, score: term, scores: term, secret: term, secrets: term}

Functions

guess(game, arg)

Specs

guess(t, Hangman.Guess.t) :: result

guess(t, Hangman.Guess.t) :: result

Guess the specified word and update the pattern state accordingly

If correct, returns the :correct_word data tuple along with game

If incorrect, returns the :incorrect_word data tuple with game

info(g)

Specs

info(t) :: Keyword.t

Returns strategy information

load(id_key, secret, max_wrong)

Specs

load(id, [String.t], pos_integer) :: t

load(id, String.t, pos_integer) :: t

Loads a new game state given new secrets

mystery_letter()

Specs

mystery_letter :: String.t

status(game)

Specs

status(t) :: tuple

Returns current game status text