Hangman v0.9.5 Hangman.Game.Event.Manager behaviour

Module implements GenStage event manager which handles event dispatching given consumer backpressure

Functions

async_notify(event)

Specs

async_notify(term) :: term

Sends an event and returns only after the event is dispatched. Blocks until event is broadcasted

handle_demand(incoming_demand, arg)

Callback implementation for c:Experimental.GenStage.handle_demand/2.

init(atom)

Callback implementation for c:Experimental.GenStage.init/1.

start_link()

Specs

start_link :: GenServer.on_start

Starts the manager.

Callbacks

handle_cast(term, term)

Specs

handle_cast(term, term) :: tuple

handle_demand(pos_integer, term)

Specs

handle_demand(pos_integer, term) :: tuple

init(term)

Specs

init(term) :: {Experimental.GenStage.type, tuple, Experimental.GenStage.options}