Hangman v0.9.5 Hangman.Game.Server behaviour

Module handles Hangman Game serving to multiple clients.
Each player's active game state is maintained, until the player process exits.

In the event the player aborts abnormally, the player's game state is maintained but removed from active status. Subsequent player access will find the game in the :abort state and the game will check if there are any games left to play properly. If so, game playing continues regularly, else, the game is properly ended with the game over results.

NOTE: The server runs each game one at a time and handles support for multiple games concurrently but is currently not harnessed by the Game.Pid.Cache.Server for initial simplicity purposes and therefore not presently utilized.

Types

id :: Hangman.Player.id

Functions

guess(game_pid, player_key, round_key, arg)

Specs

guess(pid, Hangman.Player.key, Hangman.Round.key, Hangman.Guess.t) :: map

Issues guess letter or word request, returns guess result

Internally, runs guess against game secret. Updates Hangman pattern, status, and other game recordkeeping structures.

Guesses follow two types

{:guess_letter, letter} - If correct, returns the :correct_letter data tuple along with game info otherwise, returns the :incorrect_letter data tuple along with game info

{:guess_word, word} - If correct, returns the :correct_word data tuple along with game info. If incorrect, returns the :incorrect_word data tuple with game info

register(game_pid, player_key, round_key)

Specs

register(pid, Hangman.Player.key, Hangman.Round.key) :: map

Initiates link with client and returns Game secret length

setup(game_pid, id_key, secret, max_wrong \\ 5)

Specs

setup(pid, id, String.t | [String.t], pos_integer) :: :ok

Loads new game into server process state. Used primarily by Game.Pid.Cache

start_link(id_key, secret, max_wrong \\ 5)

Specs

start_link(id, [String.t], pos_integer) :: GenServer.on_start

Start public interface method with secret(s)

status(game_pid, player_key, round_key)

Specs

status(pid, Hangman.Player.key, Hangman.Round.key) :: map

Retrieves Game status data

stop(game_pid)

Specs

stop(pid) :: tuple

Issues request to stop GenServer

whereis(id_key)

Specs

whereis(id) :: pid | :undefined

Routine returns game server pid from process registry using gproc If not found, returns :undefined

Callbacks

handle_call({}, tuple, arg2)

Specs

handle_call({atom, Hangman.Player.key, Hangman.Round.key}, tuple, Hangman.Simple.Registry.t) :: tuple

handle_call({Hangman.Guess.t, Hangman.Player.key, Hangman.Round.key}, tuple, Hangman.Simple.Registry.t) :: tuple

handle_call({Hangman.Guess.t, Hangman.Player.key, Hangman.Round.key}, tuple, Hangman.Simple.Registry.t) :: tuple

handle_call({:atom, Hangman.Player.key, Hangman.Round.key}, tuple, Hangman.Simple.Registry.t) :: tuple

handle_cast(tuple, arg1)

Specs

handle_cast(tuple, Hangman.Simple.Registry.t) :: tuple

init(term)

Specs

init(term) :: tuple