Hangman v0.9.6 Hangman.Handler.Accumulator

Module implements an accumulator which builds up a sequence list in the context of a game client handler. Hides the details of the functional reduce-while-cycle trickery

Functions

done()

done(value)

next()

next(value)

Macros

cycle(list)

Cycle builds up an accumulator sequence using next(value) until we invoke done