Hangman v0.9.6 Hangman.Ingestion.Cache.Flow

Loads partitioned dictionary word files into ingestion table piecewise through words list chunks. Random words are generated from these words lists.

Letter frequency counters of the dictionary words are arranged by length and also stored in the ingestion db after the words are stored

Optimization Note 1: Generating counter tallies outside of the ingestion table allows us to use Flow's parallelism to generate these

Optimization Note 2: Converting word_list chunks to binaries and counters to binaries drastically reduces table memory footprint

Functions

run(cache_dir, dump_path)

Specs

run(binary, binary) :: :ok

Loads up the intermediate cached files and processes the data using flow so that it can be stored in a table. The types of data generated and stored into the db are words data, random words, and letter frequency counters arranged by word key length.

Lastly, the ingestion db is dumped to file