Hangman v0.9.5 Hangman.Ingestion.First.Flow

Module implements the introductory ingestion flow logic given an initial hangman dictionary file. Serves to split dictionary into windowed word chunks storing them in intermediate cached files arranged by word key length

Functions

run(path, key_file_map)

Specs

run(binary, map) :: {:ok, map}

Run reads in dictionary words file, and checkpoints snapshots of so many events, e.g. 500 events for each partition. Each checkpoint snapshot is written to the corresponding word key length dictionary partition file. E.g. partition 0 data is written to key file 2, partition 8 data is written to key file 10. Data is written in a way so that it is easy to extract the key (word key length) and data: windowed word data