Hangman v0.9.5 Hangman.Letter.Retrieval.Common

Functions

optimal(strategy)

Returns optimal letter

Method implements the most common letter retrieval strategy with a twist. Gets the first letter with the highest frequency for when the current possible Hangman word set space is > "small". The twist is when we combine the English language letter relative frequency. For the cases where the word set is less than small, takes the letter whose frequencies are less than or equal to half the possible Hangman word pass size.

E.g.for size 10, the letter counts would need to be 5 or lower to be chosen. Doesn't handle tie between letters.