Hangman v0.9.5 Hangman.Letter.Strategy

Module provides access to a set of functions handling letter guessing strategy.

(From Wikipedia)

In the English language, the twelve most commonly occurring letters are, in descending order: e-t-a-o-i-n-s-h-r-d-l-u. This and other letter-frequency lists are used by the guessing player to increase the odds when it is their turn to guess. On the other hand, the same lists can be used by the puzzle setter to stump their opponent by choosing a word which deliberately avoids common letters (e.g. rhythm or zephyr) or one that contains rare letters (e.g. jazz).

Another common strategy is to guess vowels first, as English only has five vowels (a, e, i, o, and u, while y may sometimes, but rarely, be used as a vowel) and almost every word has at least one.

According to a 2010 study conducted by Jon McLoone for Wolfram Research, the most difficult words to guess include jazz, buzz, hajj, faff, fizz, fuzz and variations of these.

NOTE: For the implementation, letter-frequency lists are primarily used. A vowel strategy is not explicitly used, and perhaps is utilized implicitly through letter-frequency lists. No effort at this point is made to stump the player.

For the robot player type, the strategy retrieves the best letter considering English letter frequencies and letter tally counts.

For the human player type, the strategy retries a set of top letter choices to be presented to human to manually choose.

Types

result :: {t, Hangman.Guess.t}

t

Functions

choices(strategy, n \\ 5)

Specs

choices(t, pos_integer) :: Hangman.Guess.option

Retrieves letter choices text. Denotes 'strategic' letter pick with an asterisk.

guess(strategy)

Specs

guess(t) :: Hangman.Guess.t

Returns best letter guess as deemed by heuristics

guessed(strategy)

Specs

guessed(t) :: list

Returns a list of guessed letters up to this point

info(s)

Specs

info(t) :: Keyword.t

Returns Strategy information

most_common(strategy, n)

Specs

most_common(t, pos_integer) :: Keyword.t

Returns most common n letters along with their counts

new(type)

Specs

new(atom) :: t

Returns a new Letter.Strategy

update(strategy, pass)

Specs

update(t, Hangman.Guess.t) :: t

update(t, Hangman.Pass.t) :: t

Updates strategy with engine pass data.

For robot player type, when we updated the pass data also prepare the guess.

validate(strategy, letter, n \\ 5)

Specs

validate(t, pos_integer, String.t) :: Hangman.Guess.t

Validates letter is within the top strategy letter choices. If not, picks the top letter as deemed by heuristics.