Hangman v0.9.5 Hangman.Pass

Module defines types Pass.key and Pass.t

Returns result of pass runs as distinguished by initial :start or subsequent :guessing modes.

Pass data is a group of the pass size, the letter frequency tally, and relevant data on final word information.

Given a new Hangman game, initially the words pass is the size of all words in the dictionary of secret length k. As each round proceeds, this is reduced by the Hangman pattern sequence. It is these remaining possible word set instances that are stored in the cache.

After each player has made their Hangman round guess, the resultant reduced words pass data is stored into the Pass.Cache for access on the subsequent round. The expired pass data from stale rounds is subsequently removed from the cache.

Types

key :: {id :: String.t | tuple, game_num :: non_neg_integer, round_num :: non_neg_integer}

Defines word pass key type

t :: %Hangman.Pass{last_word: term, possible: term, size: term, tally: term}

Defines word pass type

Functions

cleanup(pass_key)

Specs

cleanup(key) :: :ok

increment_key(key)

Specs

increment_key(Hangman.Pass.key) :: Hangman.Pass.key

result(atom, pass_key, reduce_key)

Specs

result(atom, pass_key :: Hangman.Pass.key, reduce_key :: Hangman.Reduction.key) :: {Hangman.Pass.key, Hangman.Pass.t}

Result routine retrieves the pass size, tally, possible words, and other data given these cache keys. Relies on either the Dictionary Cache or the Reduction Engine to compute new pass data

:start - this is the initial game start pass, so we request the data from the Dictionary.Cache. The data is stored into the Pass.Cache via Pass.Cache.Writer.write/2. Returns pass data type.

:guessing - retrieves the pass data from the last player round and relies on Reduction.Engine.reduce/3 to reduce the possible Hangman words set with reduce_key. When the reduction is finished, we write the data back to the Pass.Cache and return the new pass data.