Hangman v0.9.5 Hangman.Pass.Cache

Module provides cache access to the game words pass. Given a player, a game and round number, Pass.Cache maintains a words pass cache of current word passes. The words are represented by Chunks.

Implements a get/1 and put/2 routine to retrieve and store these word pass Chunks.

Given a new Hangman game, initially the words pass is the size of all words in the dictionary of secret length k. As each round proceeds, this is reduced by the Hangman pattern sequence. It is these remaining possible word set instances that are stored in the cache.

After each player has made their Hangman round guess, the resultant reduced words pass data is stored into the Pass.Cache for access on the subsequent round. The expired pass data from stale rounds is subsequently removed from the cache.

Serves as the unified point for the reduction logic when handling data around the Pass.Cache.

Functions

get(pass_key)

Specs

get(Hangman.Pass.key) :: Hangman.Chunks.t | no_return

Get routine retrieves pass chunks cache data given the pass key

put(pass_key, data)

Specs

put(Hangman.Pass.key, Hangman.Chunks.t) ::
  :ok |
  no_return

Put routine stores new pass chunks data, provided the pass key

start_link()

stop(pid)

Specs

stop(pid) :: {}

Routine to stop server normally