Hangman v0.9.5 Hangman.Pass.Cache.Writer behaviour

Module is responsible for synchronous chunk pass writes into the Pass.Cache table.

Serves as a write operation specific process to isolate write errors from Pass.Cache

NOTE: Could be a source of bottleneck as all reduce workers are synchronously relying on this writer process. As of current load, no bottleneck for time being

Functions

put(pass_key, chunks)

Specs

put(Hangman.Pass.key, Hangman.Chunks.t) :: :ok

Write is synchronous

start_link()

Starts GenServer

Callbacks

handle_call({}, tuple, tuple)

Specs

handle_call({atom, Hangman.Pass.key, Hangman.Chunks.t}, tuple, tuple) :: tuple