Hangman v0.9.1 Hangman.Pass.Server behaviour

Module implements ets table owning process for hangman words pass state for a given player, round number.

Performs primarily unserialized reads

Functions

get_pass(atom, pass_key, reduce_key)

Specs

get_pass(atom, Hangman.Pass.key, Hangman.Reduction.key) :: {Hangman.Pass.key, Hangman.Pass.t}

get_pass(atom, Hangman.Pass.key, Hangman.Reduction.key) :: {Hangman.Pass.key, Hangman.Pass.t}

Game keep guessing engine pass routine Routine retrieves pass tally given pass key Request not serialized through server process, since we are doing reads

read_chunks(pass_key)

Specs

read_chunks(Hangman.Pass.key) ::
  Hangman.Chunks.t |
  no_return

Retrieves pass chunks from ets table with pass key”

start_link()

Specs

start_link :: Supervisor.on_start

GenServer start_link wrapper function

stop(pid)

Specs

stop(pid) :: {}

Routine to stop server normally

Callbacks

handle_call(atom, {}, {})

Specs

handle_call(:atom, {}, {}) :: {}

GenServer callback to retrieve game server pid

init(term)

Specs

init(term) :: {}

GenServer callback to initialize server process

terminate(reason, {})

Specs

terminate(reason :: term, {}) :: term | no_return

GenServer callback to cleanup server state