Elixir Hangman v0.1.1 Hangman.Pass.Server.Stub

Stub module to mimic Pass Server functionality

Provides a simple scaffold implementation to provide simple, predictable behavior

Functions

get_pass(atom, pass_key, reduce_key)

Specs

get_pass(atom, Hangman.Pass.key, Hangman.Reduction.key) :: tuple

get_pass(atom, Hangman.Pass.key, Hangman.Reduction.key) :: tuple

Stub Routine retrieves stub pass tally given pass key