Elixir Hangman v0.1.1 Hangman.Pass.Writer

Module implements words pass writer functionality. Write load is handled through pool. Distributes write request based on pass key id attribute (name) Pool size writer workers are started up as part of writer pool Pool supervisor supervises writer workers

Functions

start_link()

Specs

start_link :: Supervisor.on_start

Supervisor start_link wrapper function Starts pool supervisor

write(pass_key, chunks)

Specs

write(Hangman.Pass.key, Hangman.Chunks.t) :: :ok

Write is an asynchronous call, no need to wait around for response Based on key id, selects writer worker to hand off request to