Hangman v0.9.6 Hangman.Player

Defines Player id and key

Provides Player generic routine implementations.

For polymorphic routines, delegates to Player.Action protocol.

Player is the common DNA unifying all players.

The Player.Action protocol implement the specialized functionality of the player types

Types

id ::
  String.t |
  {id :: String.t, shard_no :: pos_integer}

key :: {id :: String.t, player_pid :: pid}

Functions

begin(player)

Begin new game player action

guess(player, guess \\ nil)

Returns player guess

new(arg)

Create new player

setup(player)

Sets up each action state

transition(player)

Returns the correct player transition at the game end