Hangman v0.9.5 Hangman.Player.Action protocol

Module implements player action functionality for various player types via protocol mechanism

The Player Action protocol is implemented for the Human and Robot types, with Generic Player handling overlaps in functionality. The Action protocol could be thought of as a sort of virtual player as it defers implementation to a combination of type specific functionality and generic type functionality.

The goal of a player is to maximize our winning chances in conjunction with a letter strategy against the 'implicit' other player, the game server.

Player action functionality handles choosing letters, guessing letters and words.

Player action encapsulates via the action types, the data used along with Strategy data. Round functionality extends the scope of the player to handle the actual game round details.

Function names are new/2, begin/1, setup/1, guess, transition/1

Types

t :: term

Functions

begin(player)

Specs

begin(term) :: term

Begin new game player action

guess(player, guess \\ nil)

Specs

guess(term, term) :: term

Returns player guess

new(player, args)

Specs

new(term, tuple) :: term

Create new player

setup(player)

Specs

setup(term) :: term

Sets up each action state

transition(player)

Specs

transition(term) :: term

Returns the correct player transition at the game end