Hangman v0.9.6 Hangman.Player.Controller

Serves as External API into Player functionality.
Serves to manage players.

Responsible for starting a player worker, and via the Player.Worker.Supervisor any worker crashes.

Forwards player requests to Player.Worker and responses back to the relevant Handler, either CLI or Web.

The Player sandwich shows the ingredient layers of the player:

cli_handler | web_handler (the player clients)

player controller (a proxy module, providing a single player interface)

player worker supervisor (dynamically starts children and handles abnormal crashes)

player worker (issues requests to fsm)

player fsm (fsm wrapper for action protocol)

player action (handles dynamic dispatch based on player types)

action human | action robot (specific types implemented)

round | strategy (handles game playing specifics -- choosing best letter, communicating with game server and reduction engine)

Functions

guess(id, data)

Specs

guess(Hangman.Player.id, tuple | String.t) :: tuple

Issues guess request with guess data

proceed(id)

Specs

proceed(Hangman.Player.id) :: tuple

Issues proceed call to iterate player sequence

start_worker(name, type, display, game_pid)

Specs

start_worker(Hangman.Player.id, atom, boolean, pid) :: :ok

Dynamically start a new player worker

stop_worker(id)

Specs

stop_worker(Hangman.Player.id) :: atom

Issues request to stop worker