Hangman v0.9.1 Hangman.Player.FSM behaviour

Module implements state machine model for player

Supports three sets of player type states: human, robot, robot turbo

Each player type is mutually exclusive and has its own state transitions

Provides a public interface with inherent asserting through specific transitions to control player behavior

Embeds player abstraction internally as state machine state

State machine is a wrapper around player abstraction

Functions

async_guess(fsm_pid)

Specs

async_guess(pid) :: :ok

Primary turbo robot event notification function used by async echo server Sends game keep guessing state transition.

code_change(OldVsn, state_name, state, extra)

Callback implementation for c::gen_fsm.code_change/4.

eager_socrates(arg1, from, arg3)

giddy_socrates(arg1, from, arg3)

handle_event(arg1, state_name, state)

Callback implementation for c::gen_fsm.handle_event/3.

handle_info(Info, state_name, state)

Callback implementation for c::gen_fsm.handle_info/3.

handle_sync_event(arg1, arg2, state_name, state)

Callback implementation for c::gen_fsm.handle_sync_event/4.

idle_socrates(arg1, from, arg3)

init(arg)

Callback implementation for c::gen_fsm.init/1.

intrigued_wall_e(atom, from, arg)

neutral_wall_e(atom, arg)

neutral_wall_e(atom, from, arg)

socrates_guess(fsm_pid, letter)

Specs

socrates_guess(pid, String.t) :: term

Sends guess_letter tuple state transition. Assumption is we are in a human / socrates state

socrates_proceed(fsm_pid)

Specs

socrates_proceed(pid) :: term

Sends proceed state transition. Assumption is we are in a human / socrates state

socrates_win(fsm_pid)

Specs

socrates_win(pid) :: term

Sends guess last word state transition. Assumption is we are in a human / socrates state

spellbound_wall_e(atom, arg)

start_link(event_server_pid, game_server_pid, player_name, player_type)

Specs

start_link(pid, pid, String.t, :atom) :: tuple

fsm start and link wrapper function

stop(fsm_pid)

Specs

stop(pid) :: :ok

Stops fsm server process

sync_games_over_status(fsm_pid)

Specs

sync_games_over_status(pid) :: term

Retrieves game over status independent of current fsm state

sync_status(fsm_pid)

Specs

sync_status(pid) :: term

Retrieves game status independent of current fsm state

terminate(reason, state_name, state)

Callback implementation for c::gen_fsm.terminate/3.

turbo_wall_e_guess(fsm_pid)

Specs

turbo_wall_e_guess(pid) :: :ok

Primary turbo robot event notification function Sends game keep guessing state transition.

wall_e_guess(fsm_pid)

Specs

wall_e_guess(pid) :: term

Primary robot event notification function Sends game keep guessing state transition.

zen_wall_e(atom, arg)

zen_wall_e(atom, from, arg)

Callbacks

eager_socrates(arg0, tuple, tuple)

Specs

eager_socrates(Guess.t, tuple, tuple) :: tuple

eager_socrates(:atom, tuple, tuple) :: tuple

eager_socrates(:atom, tuple, tuple) :: tuple

Callback function for synchronous human state socrates and event tuple :guess_letter

giddy_socrates(arg0, tuple, tuple)

Specs

giddy_socrates(Guess.directive, tuple, tuple) :: tuple

giddy_socrates(tuple, tuple, tuple) :: tuple

giddy_socrates(:atom, tuple, tuple) :: tuple

Callback function for synchronous human state socrates and event :guess_last_word

handle_sync_event(atom, tuple, atom, tuple)

Specs

handle_sync_event(:atom, tuple, :atom, tuple) :: tuple

handle_sync_event(:atom, tuple, :atom, tuple) :: tuple

Callback provides state-independent game round status reply

idle_socrates(atom, tuple, tuple)

Specs

idle_socrates(:atom, tuple, tuple) :: tuple

idle_socrates(:atom, tuple, tuple) :: tuple

idle_socrates(tuple, tuple, tuple) :: tuple

Callback function for synchronous human state socrates and event :proceed

init(tuple)

Specs

init(tuple) :: {:ok, :atom, tuple}

State machine initial callback routine

Loads a new player abstraction and echo server, returns first game state

intrigued_wall_e(atom, tuple, tuple)

Specs

intrigued_wall_e(:atom, tuple, tuple) :: tuple

Callback function for synchronous robot state wall_e and event :game_keep_guessing for when game is progressing and player is guessing

neutral_wall_e(atom, tuple)

Specs

neutral_wall_e(:atom, tuple) :: tuple

Callback function for asynchronous robot state wall_e and event :game_keep_guessing for when game is just starting or all games over

neutral_wall_e(atom, tuple, tuple)

Specs

neutral_wall_e(:atom, tuple, tuple) :: tuple

Callback function for synchronous robot state wall_e and event :game_keep_guessing for when game is just starting or just over

spellbound_wall_e(atom, tuple)

Specs

spellbound_wall_e(:atom, tuple) :: tuple

Callback function for asynchronous robot state wall_e and event :game_keep_guessing for when game is progressing and robotic player has started guessing

terminate(term, atom, term)

Specs

terminate(term, :atom, term) :: term

State machine terminate function callback

zen_wall_e(atom, tuple)

Specs

zen_wall_e(:atom, tuple) :: tuple

Callback function for asynchronous robot state wall_e and event :game_keep_guessing for when all games are over

Returns stop gen fsm reply

zen_wall_e(atom, tuple, tuple)

Specs

zen_wall_e(:atom, tuple, tuple) :: tuple

Callback function for synchronous robot state wall_e and event :game_keep_guessing for when all games are over