Hangman v0.9.1 Hangman.Player.Round

Module to implement player game round abstraction.

Works in conjuction with Strategy to orchestrate actual round game play.

Basic round functionality includes as setup, guess, update, status

Types

result :: {status_code :: :atom, status_txt :: String.t}

t :: %Hangman.Player.Round{final_result: term, guess: term, pattern: term, result_code: term, seq_no: term, status_code: term, status_text: term}

Functions

augment_choices(player, arg)

Specs

augment_choices(Hangman.Player.t, Hangman.Guess.option) :: Hangman.Guess.option

Interjects round specific parameters into choices text

context(player)

Specs

context(Hangman.Player.t) ::
  Hangman.Guess.context |
  no_return

Returns round context based on results of last guess

guess(p, arg)

Specs

guess(Hangman.Player.t, Hangman.Guess.t) :: t

guess(Hangman.Player.t, Hangman.Guess.t) :: t

Issues a client guess (either letter or word) against Game Server. Notifies player events server of guess results.

Returns received round data

setup(player)

Specs

setup(Hangman.Player.t) :: Hangman.Player.t

Setups game play round

For game start stage, retrieves secret length from game server uses secret length to filter possible hangman words from pass server

On subsequent rounds, generates a reduce key based on the result of the last guess to filter possible hangman words from word pass server

setup(player, context)

Specs

setup(Hangman.Player.t, Hangman.Guess.context) :: Hangman.Player.t

setup(Hangman.Player.t, :atom) :: Hangman.Player.t

status(player)

Specs

status(Hangman.Player.t) :: result

Returns round status tuple

update(player, round)

Specs

update(Hangman.Player.t, Hangman.Player.Round.t) :: Hangman.Player.t

update(player, round, arg)

Specs

update(Hangman.Player.t, Hangman.Player.Round.t, Hangman.Guess.t) :: Hangman.Player.t

update(Hangman.Player.t, Hangman.Player.Round.t, Hangman.Guess.t) :: Hangman.Player.t

Updates player abstraction with round results. If games are over, updates games summary and notifies player events server.

Under human guessing, player round update will update the strategy abstraction with the guess particulars. If robot guessing, the strategy abstraction will also be updated.