Hangman v0.9.5 Hangman.Player.Worker behaviour

GenServer module to implement Player Worker, managed by Player.Controller. Module uses gproc to manage process registry of workers.

The module represents the highest effective player worker abstraction.

It sits in conjunction with other player components, between the Game and Reduction engines and the Client.handler - a producer-consumer.

Behind this GenServer lies the intermediary player components which facilitate player game play. These are Player.Action, Player.Human, Player.Robot, Player.Generic, Player.FSM, Round, Strategy.

Internally the Worker keeps a Player.FSM as a state to manage event transitions smoothly.

The module is abstracted away from the specific type of player to focus mainly on feeding the Player FSM and returning the appropriate response to the Player.Controller.

The Player sandwich shows the ingredient layers of the player:

cli_handler | web_handler (the player clients)

player controller (a proxy module, providing a single player interface)

player worker supervisor (dynamically starts children and handles abnormal crashes)

player worker (issues requests to fsm)

player fsm (fsm wrapper for action protocol)

player action (handles dynamic dispatch based on player types)

action human | action robot (specific types implemented)

round | strategy (handles game playing specifics -- choosing best letter, communicating with game server and reduction engine)

The player worker interface constitutes of two methods: proceed/1 and guess/2

Functions

guess(worker_id, data)

Specs

guess(Hangman.Player.id, tuple) :: tuple

guess(Hangman.Player.id, String.t) :: tuple

proceed(worker_id)

Specs

proceed(Hangman.Player.id) :: tuple

start_link(arg)

stop(worker_id)

Specs

stop(Hangman.Player.id) :: atom

whereis(worker_id)

Specs

whereis(Hangman.Player.id) :: pid | :atom

Routine returns game server pid from process registry using gproc If not found, returns :undefined

Callbacks

handle_call(atom, tuple, any)

Specs

handle_call(:atom, tuple, any) :: tuple

handle_call(tuple, tuple, term) :: tuple

handle_call(tuple, tuple, term) :: tuple

handle_call(atom, tuple, term) :: tuple

init(term)

Specs

init(term) :: tuple

Starts up new FSM and initializes it with worker args

terminate(term, term)

Specs

terminate(term, term) :: :ok

Terminate callback.