Hangman v0.9.5 Hangman.Reduction.Engine

Module provides access to the words reduction engine.

Reduces possible Hangman words set based on the provided reduce key. Reduction load is handled through Reduction.Engine.Pool.

The Reduction.Engine distributes reduce/3 requests based on the pass key id attribute to workers. Engine workers are started up as part of the reducer pool.

The pool supervises the reduction workers.

Functions

reduce(pass_key, regex_key, exclusion_set)

Specs

reduce(Hangman.Pass.key, Regex.t, Enumerable.t) :: Hangman.Pass.t

Distributes reduce request to appropriate worker. Calls synchronous worker process function Reduction.Engine.Worker.reduce_and_store/4. Hands off request based on key id.

start_link()

Specs

start_link :: Supervisor.on_start

Supervisor start link wrapper function Starts pool Supervisor