Hangman v0.9.5 Hangman.Round

Module provides access to game round functions.

Round represents the time period in the Hangman game which consists of setup steps: word set reduction, guess assistance.

It works in conjuction with Strategy and Game.Server to orchestrate actual round game play through guess actions.

When playing a Hangman game, we first init our round, which involves obtaining the secret word length from the game server.

Next, we take steps to reduce the possible Hangman words set to narrow our word choices.

From there on we choose the best letter to guess given the knowledge we have of our words data set. If we are a :human, we are given letter choices to pick, and if we are a :robot, we trust our friend Strategy.

After a guess is made either by :human or :robot we update our round recordkeeping structures with the guess results and proceed for the next round -- to do it all over again.

Basic Round functionality includes init/1, setup/3, guess/2, transition/1, status/1.

We always invoke setup before a guess, to properly setup the new words state.

Transition is used after a single game over to determine if we transition to a new game or games over.

Types

context ::
  {:start, secret_length :: pos_integer} |
  {:guessing, :correct_letter, letter :: String.t, pattern :: String.t, mystery_letter :: String.t} |
  {:guessing, :incorrect_word, word :: String.t} |
  {:guessing, :incorrect_letter, letter :: String.t} |
  {:won, :correct_word, word :: String.t}

Sum type used to understand prior guess result

key :: {id :: String.t, game_num :: integer, round_num :: integer}

result_code ::
  :correct_letter |
  :incorrect_letter |
  :incorrect_word

t :: %Hangman.Round{context: term, game_num: term, game_pid: term, guess: term, id: term, num: term, pattern: term, pid: term, result_code: term, status_code: term, status_text: term}

Functions

guess(round, arg)

Specs

guess(t, Hangman.Guess.t) :: t

Issues a client guess (either letter or word) against Game.Server. Returns received round data

info(round)

Specs

info(t) :: Keyword.t

init(round)

Specs

init(t) :: t

Initialize the round with the start of a new game. Retrieves the secret length from the game server and creates a process link to the game server via the register call.

Eventually the secret length is used to filter possible Hangman words from Pass.Cache server on the next round setup.

round_key(round)

Specs

round_key(t) :: tuple

Returns round tuple key

setup(round, exclusion, fn_updater)

Specs

setup(Hangman.Round.t, List.t, (Map.t -> Hangman.Letter.Strategy.t)) :: Hangman.Round.t

Sets up game play round

Generates a reduce key based on the result of the last guess or secret length to filter possible Hangman words from Pass.Cache server

status(round)

Specs

status(t) :: {Hangman.Game.Server.code, String.t}

Returns round status tuple

transition(round)

Specs

transition(t) :: t

Specifies steps for the end of a single game. Transitions to either a :start or :finished depending on if there are games left to play