Hangman v0.9.6 Hangman.Shard.Handler

Module runs game play for the given shard of secrets as determined by Shard.Flow. Basically, runs a chunk of the overall original secrets vector.

Module drives Player.Controller, while setting up the proper Game server and Event consumer states beforehand.

Simply stated it politely nudges the player to proceed to the next course of action or make the next guess.

When the game is finished it politely ends the game playing returning the shard_key and game snapshot tuple.

The twist to this is that these shard handlers are run in parallel and concurrently thanks to the concurrent map reduce setup of Flow

Functions

play(shard_key)

Specs

play(Hangman.Player.id) :: {Hangman.Player.id, [String.t]}

Play handles client play loop for particular player shard_key

setup(arg)

Specs

setup({Hangman.Player.id, list[<a href="http://elixir-lang.org/docs/stable/elixir/String.html#t:t/0">String.t</a>]}) :: Hangman.Player.id

Sets up the game server and per player event server. Used primarly by the collation logic in Flow