Hangman v0.9.1 Hangman.Strategy.Options

Module generates reduction key for use when reducing possible words set

Used primarily by player round abstraction before guessing, when setting up a round

Functions

reduce_key(strategy, context)

Specs

reduce_key(Hangman.Strategy.t, Hangman.Guess.context) :: Hangman.Reduction.key

reduce_key(Hangman.Strategy.t, Hangman.Guess.context) :: Hangman.Reduction.key

reduce_key(Hangman.Strategy.t, Hangman.Guess.context) :: Hangman.Reduction.key

Generates reduction key given round context

regex_match_key(arg, guessed_letters)

Specs

regex_match_key(Hangman.Guess.context, map) :: Regex.t

regex_match_key(Hangman.Guess.context, map) :: Regex.t

Generates regex key to match against possible hangman words

For incorrect letter last guesses, uses the fact the we know the incorrect letter can not be found anywhere in the possible hangman words.

We create a regex to reflect this information