Hangman v0.9.5 Hangman.Web

Module provides access to a http web server for playing Hangman games via a tool such as curl or HTTPoison.

Query params are specified after /hangman?

These are name :: String.t, secret :: String.t | [String.t], random :: pos_integer

The name at all times must be specified along with either a secret value or random value: true = name and (secret | random)

The random param specifies how many secret words to generate. Hence, random=5, will generate 5 secret game words for a total of 5 hangman games.

If the secrets collection vector is only 1 item we show the game history if greater we show the word and score summary only

Example

iex> HTTPoison.get("http://127.0.0.1:3737/hangman?name=julio&secret=kiwi")

{:ok,
%HTTPoison.Response{body: "(H) ----; score=1; status=KEEP_GUESSING (H) ----; score=2; status=KEEP_GUESSING (H) ----; score=3; status=KEEP_GUESSING (H) -I-I; score=4; status=KEEP_GUESSING (H) -I-I; score=5; status=KEEP_GUESSING (H) -I-I; score=6; status=KEEP_GUESSING (H) -I-I; score=25; status=GAME_LOST (H) Game Over! Average Score: 25.0, Games: 1, Scores:  (KIWI: 25) ",
headers: [{"server", "Cowboy"}, {"date", "Mon, 14 Mar 2016 04:27:39 GMT"},
{"content-length", "345"},
{"cache-control", "max-age=0, private, must-revalidate"},
{"content-type", "text/plain; charset=utf-8"}], status_code: 200}}

iex> HTTPoison.get("http://127.0.0.1:3737/hangman?name=julio&random=1")

{:ok,
%HTTPoison.Response{body: "(H) --E-----------; score=1; status=KEEP_GUESSING (H) --E--------O--; score=2; status=KEEP_GUESSING (H) --E--------O--; score=3; status=KEEP_GUESSING (H) --E----C---O--; score=4; status=KEEP_GUESSING (H) PREVARICATIONS; score=4; status=GAME_WON (H) Game Over! Average Score: 4, Games: 1, Scores:  (PREVARICATIONS: 4) ",
headers: [{"server", "Cowboy"}, {"date", "Mon, 14 Mar 2016 04:29:05 GMT"},
{"content-length", "601"},
{"cache-control", "max-age=0, private, must-revalidate"},
{"content-type", "text/plain; charset=utf-8"}], status_code: 200}}

Functions

call(conn, opts)

Callback implementation for c:Plug.call/2.

init(opts)

Callback implementation for c:Plug.init/1.

run(conn)

Specs

run(Plug.Conn.t) :: Plug.Conn.t | no_return

Retrieves connection params name and either a secret or random. Runs web game. Returns complete game results in connection response body

start_server()

Starts the cowboy web server