Hangman v0.9.5 Hangman.Web.Shard.Handler

Module runs game play for the given web shard of secrets as determined by Web.Flow. Basically, runs a chunk of the overall original secrets vector.

Module drives Player.Controller, while setting up the proper Game server and Event consumer states beforehand.

Simply stated it politely nudges the player to proceed to the next course of action or make the next guess. The handler also collects input from the user as necessary and displays data back to the user.

When the game is finished it politely ends the game playing.

Functions

play(player_handler_key)

Specs

play(Hangman.Player.id) :: tuple

Play handles client play loop

setup(arg)

Specs

setup(tuple) :: tuple

Sets up the game server and per player event server. Used primarly by the Web.Collator