Hangman v0.9.5 Hangman.Words

Module to manage word lists for a given secret length key.

The need for splitting lists arises when we have arbitrary long word lists/streams, as it provides structure.

Module internally maintains standardized containers for word lists and keeps track of total word counts. Splits big words list into smaller lists.

A single Words abstraction can contain a single word list or multiple lists with sizes info, which is represented as a container.

Words provide extra manageability during storage as lists are binaried leaving a smaller footprint.

Primary functions are new/2, count/1, add/2, filter/2, and stream/1.

Types

t

Functions

add(words, value)

Specs

add(t, {[binary], integer}) :: t

Takes an existing Words and adds the passed in container tuple.

Used in reduce to add a {[binary], count} to the Words accumulator value.

The tuple head is a binaried word list and the tail is the word count

collect(words, count)

Specs

collect(t, pos_integer) :: String.t | [String.t]

Return specified number of words from Words

container_size()

count(words)

Specs

count(t) :: integer

Returns total word count across all Words containers

filter(words, regex_filter_key)

Specs

filter(t, Regex.t) :: Enumerable.t

Return words filtering out words which don't match the regex key, returning a new Words abstraction

info(words)

Specs

info(t) :: Keyword.t

key(words)

Specs

key(t) :: pos_integer

Returns Words word length key

new(length_key)

Specs

new(pos_integer) :: t

Returns new empty Words abstraction

new(length_key, words)

Specs

new(pos_integer, Enumerable.t) :: t

Returns new Words abstraction. Does this by splitting and encapsulating words lists from enumerable into standardized containers.
Word lists are binaried for compactness.

stream(words)

Specs

stream(t) :: Enumerable.t

Returns words in lazy enumerable fashion.