Hangman v0.9.1 Pass.Cache

Given a player, a game and round number, Pass.Cache maintains a words pass cache.

Internally the module implements a GenServer which uses ETS.

After each player has made their Hangman round guess, the resultant reduced words pass data is stored into the Pass.Cache for access on the subsequent round. The expired pass data from stale rounds is subsequently removed from the cache.

Pass.Cache performs unserialized reads and uses type key for cache get/2 and get/3.

Types

key ::
  :chunks |
  {:pass, :game_start} |
  {:pass, :game_keep_guessing}

Functions

get(cache_key, pass_key)

Specs

get(Pass.Cache.key, Pass.key) ::
  Chunks.t |
  no_return

Get routine retrieves pass chunks cache data

We can obtain chunks data, for cache keys with the :chunks atom * :chunks - retrieves chunks data for a given pass key

These get requests are not serialized through the server process since we are doing reads

get(cache_key, pass_key, reduce_key)

Specs

get(cache_key :: Pass.Cache.key, pass_key :: Pass.key, reduce_key :: Reduction.key) ::
  {Pass.key, Pass.t} |
  no_return

Get routine retrieves the pass size, tally, possible words, and other data given these cache keys.

{:pass, :game_start} - this is the initial game start pass, so we request the data from the Dictionary.Cache. The data is stored into the Pass.Cache via Pass.Writer.write/2. Returns pass data type.

{:pass, :game_keep_guessing} - retrieves the pass data from the last player round and relies on Reduction.Engine.reduce/3 to reduce the possible Hangman words set with reduce_key. When the reduction is finished, we write the data back to the Pass.Cache and return the new pass data.

These get requests are not serialized through the server process since we are doing reads

start_link()

stop(pid)

Specs

stop(pid) :: {}

Routine to stop server normally