Hangman v0.9.1 Pass.Writer

Module is responsible for words pass writes into the Pass.Cache table. The write load is handled through the Pass.Writer.Pool. It distributes write/2 request based on the pass key id attribute to workers.

The writer pool supervisor supervises the workers, which are each responsible for the write operations. The primary writer worker method is Pass.Writer.Worker.write/3

Functions

start_link()

Specs

start_link :: Supervisor.on_start

Supervisor start_link wrapper function Starts pool supervisor

write(pass_key, chunks)

Specs

write(Pass.key, Chunks.t) :: :ok

Write is an asynchronous call, there is no need to wait around for the response. Based on the key id, selects the writer worker to hand the request off to.