Hangman v0.9.1 Pass.Writer.Worker behaviour

Module is a GenServer that implements writer worker functionality. Specifically, module is a write operation specific module that performs async write into pass ets table.

If write operation fails for whatever reason, doesn’t bring down the reads and table-owning pass ets owning process. Hence the separation.

Functions

start_link(worker_id)

Specs

start_link(pos_integer) :: GenServer.on_start

GenServer start_link wrapper function

stop(pid)

Specs

stop(pid) :: {}

Issues request to stop GenServer worker

write(worker_id, pass_key, chunks)

Specs

write(pos_integer, Pass.key, Chunks.t) :: :ok

Write is an asynchronous call. Insert chunks into ets pass table

Callbacks

handle_call(atom, tuple, {})

Specs

handle_call(:atom, tuple, {}) :: tuple

Issues request to stop GenServer

handle_cast(atom, arg1, arg2, {})

Specs

handle_cast(:atom, Pass.key, Chunks.t, {}) :: tuple

Writes pass data chunk, uses pass_key for ets insertion

init({})

Specs

init({}) :: {}

GenServer callback to initalize server process

terminate(term, term)

Specs

terminate(term, term) :: :ok

Terminates the pass writer worker server No special cleanup