Hangman v0.9.1 Pattern

Module handles Hangman pattern updates given guessed letter.

Given an initial starting state of ------- and given a letter a, it will update the pattern to a---a--, for the secret 'avocado'.

After letters o, and d, we are two letters shy of avocado.

Example

iex> p = Pattern.update "-------", "avocado",  "a"
"a---a--"

iex> p = Pattern.update p, "avocado",  "o"        
"a-o-a-o"

iex> p = Pattern.update p, "avocado",  "d"
"a-o-ado"

Functions

update(pattern, secret, letter)

Specs

update(String.t, String.t, String.t) ::
  String.t |
  no_return

Tail recursive function, returns pattern string after update