Hangman v0.9.1 Player

Module handles Player abstraction and defines Player type. Implements player functionality to choose letters, guess letters and words. Heavily relies upon Round and Action.

Types

kind :: :human | :robot

result :: {t, Round.result}

t

Functions

choices(p, directive, options \\ nil)

Specs

choices(t, Guess.directive, :atom) :: {t, Guess.option}

Routine for :human player type. Setups new round, retrieves and returns top letter choices.

delete(p)

Specs

delete(t) :: t

Method returns empty Player

game_lost?(p)

Specs

game_lost?(t) :: boolean

Returns true or false whether the game has been lost

game_won?(p)

Specs

game_won?(t) :: boolean

Returns true or false whether the game has been won

games_over?(p)

Specs

games_over?(t) :: boolean

Returns true or false whether all games are over

games_summary(args)

Specs

games_summary(Keyword.t) :: String.t

Returns game summary as a string. Includes number of games played, average score per game, per game score.

guess(p)

Routine for :robot player type. Setups new round, performs auto-generated guess, returns round status

guess(p, guess)

Specs

guess(p :: t, guess :: Guess.directive | Guess.t) :: result

Routine for :human player type.

Comes in two modes

:guess_last_word - performs guess of last remaining word Note: Somewhat of an oversimplification for a human guess word, we simplify the human guessing of words to just the last word

{:guess_letter, letter} - doesn’t setup round since it was already setup during the choose letters stage. Issues action to guess and validate letter and returns round status.

human()

info(p)

Specs

info(t) :: Keyword.t

last_word?(p)

Specs

last_word?(t) :: boolean

Returns true or false whether we’ve arrived at the last word in possible words set

new(name, type, game_pid, event_pid)

Specs

new(String.t, :atom, pid, pid) :: t

Returns new Player, validates player type

robot()

start(p)

Specs

start(t) :: result

Routine starts a new Player. Notifies player specific event server. Setups game round.

If type is robot, makes initial guess, then returns round status. If type is human, retrieves top letter choices to display

status(p, atom)

Specs

status(t, :atom) :: Round.result

Returns game status