Hangman v0.9.1 Player.Async.Echo behaviour

Module serves as an asynchronous echo server for the asynchronous robot fsm player type

Allows the asynchronous robot fsm player to run a game asynchronously without having to be wrapped in a Stream.run, runs automatically itself

Functions

code_change(OldVsn, state_name, state, extra)

Callback implementation for c::gen_fsm.code_change/4.

echo(arg, state)

echo_guess(fsm_pid, other_pid)

Specs

echo_guess(pid, pid) :: {}

public api method to send an echo in the form of a async guess call back to the Player.FSM

handle_event(event, state_name, state)

Callback implementation for c::gen_fsm.handle_event/3.

handle_info(arg1, state_name, state)

Callback implementation for c::gen_fsm.handle_info/3.

handle_sync_event(event, _from, state_name, state)

Callback implementation for c::gen_fsm.handle_sync_event/4.

init()

Callback implementation for c::gen_fsm.init/1.

start_link()

Specs

start_link :: {:ok, pid}

gen fsm start_link wrapper function

terminate(reason, state_name, state)

Callback implementation for c::gen_fsm.terminate/3.

Callbacks

echo({}, list)

Specs

echo({:atom, pid}, []) :: {:atom, :atom, []}

gen fsm callback for initial and only state echo handles :echo_guess event

init(arg0)

Specs

init(Keyword.t) :: {:atom, :atom, []}

gen fsm callback to initalize server process