Hangman v0.9.1 Player.Events

Module handles player event manager.

Provides access to write event notification data to a player log file or to a displayed user feed.

Functions

notify_games_over(pid, name, text)

Specs

notify_games_over(pid, name :: String.t, text :: String.t) :: :ok

Sends :games_over tuple event notification to Player.Events

notify_guess(pid, guess, arg)

Specs

notify_guess(pid, guess :: Guess.t, info :: tuple) :: :ok

Sends either a {:guess_letter, letter} or {:guess_word, word} tuple event notification to Player.Events, depending on guess.

notify_length(pid, arg)

Specs

notify_length(pid, data :: tuple) :: :ok

Sends :secret_length tuple event notification to Player.Events

notify_start(pid, name)

Specs

notify_start(pid, name :: String.t) :: :ok

Sends :start tuple event notification to Player.Events

notify_status(pid, arg)

Specs

notify_status(pid, data :: tuple) :: :ok

Sends :round_status tuple event notification to Player.Events

start_link(options \\ [file_output: false, display_output: true])

Specs

start_link(Keyword.t) :: {:ok, pid}

Start public interface method

stop(pid)

Specs

stop(pid) :: :ok

Issues request to stop GenEvent manager

terminate(reason, state)