Hangman v0.9.1 Player.Game

Module handles game playing for synchronous human and robot player types. Handles relationship between Player.FSM, Game.Server and Player.Events

Loads up the player specific game components: a dynamic game server, and a dynamic event server

Manages specific player fsm behaviour (human or robot). Wraps fsm game play into an enumerable for easy running.

Functions

rounds_handler(arg, atom)

Specs

rounds_handler(tuple, Player.kind) :: Enumerable.t

Permits robot or human round playing! Wraps the robot or human player game playing in a stream. Stream resource returns an enumerable.

Terminates player events server and fsm upon finish

run(name, type, secrets, log, display)

Specs

run(String.t, Player.kind, [String.t], boolean, boolean) :: :ok

run0(name, type, secrets, log, display)

Specs

run0(String.t, Player.kind, [String.t], boolean, boolean) :: :ok

Function run connects all the player specific components together and runs the player game

setup(name, secrets, log, display)

Specs

setup(String.t, [String.t], boolean, boolean) :: tuple

Function setup loads the player specific game components. Setups the game server and per player event server.

start_player(name, type, game_pid, notify_pid)

Specs

start_player(String.t, Player.kind, pid, pid) :: Supervisor.on_start_child

Start dynamic player child worker