Hangman v0.9.1 Player.Logger.Handler behaviour

Module implements event logger handler for player events.

Each event is logged to a file named after player id (name).

Callbacks

handle_event(tuple, term)

Specs

handle_event(tuple, term) :: tuple

The handle_event callback handles various events

:start notification event. Opens and writes to player logger file

:games_over notification event. Writes to player logger file and then closes file

:secret_length notification event. Writes to player logger file

:guessed_word notification event. Writes to player logger file

:guessed_letter notification event. Writes to player logger file

:round_status notification event. Writes to player logger file

init(term)

Specs

init(term) :: tuple

terminate(term, pid)

Specs

terminate(term, pid) :: :ok | tuple

Terminate callback. Closes player log file