Hangman v0.9.1 Reduction.Engine

Module implements words reduction engine.

Reduction load is handled through pool.

Module distributes reduce requests based on pass key id attribute (name). Reduction workers are started up as part of the reducer pool.

Pool supervises reduction workers.

Functions

reduce(pass_key, regex_key, exclusion_set)

Specs

reduce(Pass.key, Regex.t, map) :: Pass.t

Calls synchronous worker process function reduce and store Based on key id, selects reduction worker to hand off request to

start_link()

Specs

start_link :: Supervisor.on_start

Supervisor start_link wrapper function Starts pool supervisor