Hangman v0.9.1 Strategy

Module provides access to a set of functions handling guessing strategy.

From Wikipedia

In the English language, the twelve most commonly occurring letters are, in descending order: e-t-a-o-i-n-s-h-r-d-l-u. This and other letter-frequency lists are used by the guessing player to increase the odds when it is their turn to guess. On the other hand, the same lists can be used by the puzzle setter to stump their opponent by choosing a word which deliberately avoids common letters (e.g. rhythm or zephyr) or one that contains rare letters (e.g. jazz).

Another common strategy is to guess vowels first, as English only has five vowels (a, e, i, o, and u, while y may sometimes, but rarely, be used as a vowel) and almost every word has at least one.

According to a 2010 study conducted by Jon McLoone for Wolfram Research, the most difficult words to guess include jazz, buzz, hajj, faff, fizz, fuzz and variations of these.

NOTE: For the implementation, letter-frequency lists are primarily used. A vowel strategy is not explicitly used, and perhaps is utilized implicitly through letter-frequency lists. No effort at this point is made to stump the player.

For the robot player type, the strategy retrieves the best letter considering English letter frequencies and letter tally counts.

For the human player type, the strategy retries a set of top letter choices to be presented to human to manually choose.

Types

result :: {t, Guess.t}

t

Functions

choose_letters(strategy, n \\ 5)

Specs

choose_letters(t, pos_integer) :: Guess.option

Retrieves letter choices text. Denotes 'strategic' letter pick with an asterisk.

get_guessed(strategy)

Specs

get_guessed(t) :: list

Returns a list of guessed letters up to this point

info(s)

Specs

info(t) :: Keyword.t

Returns Strategy information

last_word(strategy)

Specs

last_word(t) :: Guess.t

If we have reached the last possible hangman word, return it so we can guess it

letter_in_most_common(strategy, letter, n \\ 5)

Specs

letter_in_most_common(t, pos_integer, String.t) :: Guess.t

Validates letter is within the top strategy letter choices. If not, picks the top letter as deemed by heuristics.

make_guess(strategy)

Specs

make_guess(t) :: Guess.t

Returns best letter guess as deemed by heuristics

most_common_letter(strategy, n)

Specs

most_common_letter(t, pos_integer) :: list

Returns most common n letters

most_common_letter_and_counts(strategy, n)

Specs

most_common_letter_and_counts(t, pos_integer) :: Keyword.t

Returns most common n letters along with their counts

new()

Specs

new :: t

Returns a new Strategy

retrieve_best_letter(strategy)

Specs

retrieve_best_letter(t) :: String.t

Method implements the most common letter retrieval strategy with a twist. Gets the first letter with the highest frequency for when the current possible Hangman word set space is > "small". The twist is when we combine the English language letter relative frequency. For the cases where the word set is less than small, takes the letter whose frequencies are less than or equal to half the possible Hangman word pass size.

E.g.for size 10, the letter counts would need to be 5 or lower to be chosen. Doesn't handle tie between letters.

update(strategy, arg)

Specs

update(t, Guess.t) :: t

Updates strategy with guess particulars

update(strategy, pass, atom)

Specs

update(t, Pass.t, Player.kind) :: t

Updates strategy with pass data.

For robot player type, when we updated the pass data also prepare the guess.