Hangman v0.9.1 Web

Module provides access to a http web server for playing Hangman games via a tool such as curl or HTTPoison.

Example

iex> HTTPoison.get("http://127.0.0.1:3737/play?name=julio&secret=kiwi")

{:ok,
%HTTPoison.Response{body: "(#) ----; score=1; status=KEEP_GUESSING (#) ----; score=2; status=KEEP_GUESSING (#) ----; score=3; status=KEEP_GUESSING (#) -I-I; score=4; status=KEEP_GUESSING (#) -I-I; score=5; status=KEEP_GUESSING (#) -I-I; score=6; status=KEEP_GUESSING (#) -I-I; score=25; status=GAME_LOST (#) Game Over! Average Score: 25.0, # Games: 1, Scores:  (KIWI: 25) ",
headers: [{"server", "Cowboy"}, {"date", "Mon, 14 Mar 2016 04:27:39 GMT"},
{"content-length", "345"},
{"cache-control", "max-age=0, private, must-revalidate"},
{"content-type", "text/plain; charset=utf-8"}], status_code: 200}}

iex> HTTPoison.get("http://127.0.0.1:3737/play?name=julio&random=2")

{:ok,
%HTTPoison.Response{body: "(#) --E-----------; score=1; status=KEEP_GUESSING (#) --E--------O--; score=2; status=KEEP_GUESSING (#) --E--------O--; score=3; status=KEEP_GUESSING (#) --E----C---O--; score=4; status=KEEP_GUESSING (#) PREVARICATIONS; score=4; status=GAME_WON (#) ---------; score=1; status=KEEP_GUESSING (#) --A------; score=2; status=KEEP_GUESSING (#) -IA----I-; score=3; status=KEEP_GUESSING (#) -IA----I-; score=4; status=KEEP_GUESSING (#) -IAG---I-; score=5; status=KEEP_GUESSING (#) DIAGNOSIS; score=5; status=GAME_WON (#) Game Over! Average Score: 4.5, # Games: 2, Scores:  (PREVARICATIONS: 4) (DIAGNOSIS: 5) ",
headers: [{"server", "Cowboy"}, {"date", "Mon, 14 Mar 2016 04:29:05 GMT"},
{"content-length", "601"},
{"cache-control", "max-age=0, private, must-revalidate"},
{"content-type", "text/plain; charset=utf-8"}], status_code: 200}}

Functions

call(conn, opts)

Callback implementation for c:Plug.call/2.

init(opts)

Callback implementation for c:Plug.init/1.

run_game(conn)

Specs

run_game(Plug.Conn.t) :: Plug.Conn.t

Retrieves connection params name and either a secret or random. Runs web game. Returns complete game results in connection response body

start_server()

Starts the cowboy web server