Hangman v0.9.1 Dictionary

Module defines dictionary common attributes and types functionality.

Serves as a central point to access or update such attributes

Types

kind :: :regular | :big

transform :: :unsorted | :sorted | :grouped | :chunked

Functions

big()

Specs

big :: kind

Returns big dictionary file type

chunked()

Specs

chunked :: transform

Returns chunked dictionary file type

chunks_file_delimiter()

Delimiter token used to delimit chunk values in binary chunks file

grouped()

Specs

grouped :: transform

Returns grouped dictionary file type

paths()

Returns dictionary file paths map, arranged by dictionary file types regular and big

regular()

Specs

regular :: kind

Returns regular dictionary file type

sorted()

Specs

sorted :: transform

Returns sorted dictionary file type

unsorted()

Specs

unsorted :: transform

Returns unsorted dictionary file type