Hangman v0.9.1 Dictionary.Cache

Module loads Dictionary.File words into chunks and stores into ETS. Letter frequency tallies are computed and stored into ETS upon startup. Words identified as random are tagged and stored.

Implements GenServer.

Provides lookup routines to access chunks, tallys, and random words

Functions

lookup(atom, count)

Specs

lookup(mode :: atom, pos_integer) ::
  Chunks.t |
  Counter.t |
  [String.t] |
  no_return

Cache lookup routines

The allowed modes:

:random - extracts count number of random hangman words.

:tally - retrieve letter tally associated with word length key

:chunk - retrieve the word data chunk associated with the word length key

setup(args)

Specs

setup(Keyword.t) :: :ok | no_return

Loads normalized Dictionary.File into ETS. Calculates and loads chunked word lists. Computes and stores letter tallies by word length key. Tags and stores random words.

start_link(args \\ [regular: true])

Specs

start_link(Keyword.t) :: {:ok, pid}

GenServer start link wrapper function

stop()

stop(pid)

Specs

stop(none | pid) :: {}

Routine to stop server normally