Hangman v0.9.1 Dictionary.File

Provides abstraction for various dictionary files.

The original dictionary file is transformed into intermediate files as representations of each transformation layer.

Layers range from unsorted to sorted, sorted to grouped, and grouped to chunked

Each transform handler encapsulates each layers transform procedure

Functions

transform(path, arg, dict_type)

Specs

transform(String.t, {:atom, :atom}, :atom) :: String.t

Takes input file, applies a transform type and returns new file path For example, can be used to first sort a file, then upon second invocation, group that file