Hangman v0.9.1 Game

Module defines Hangman Game abstraction which can handle a single Hangman game or multiple Hangman games.

Runs each game or games sequentially. Therefore only one game is in play at any one time.

Primary functions are Game.load/3, Game.guess/2, Game.status/1.

Types

id :: String.t

result :: {t, tuple}

t

Functions

guess(game, arg)

Specs

guess(t, guess :: Guess.t) :: result

Runs guess against Game secret. Updates Hangman pattern, status, and other game recordkeeping structures.

Guesses follow two types

{:guess_letter, letter} - If correct, returns the :correct_letter data tuple along with game data otherwise, returns the :incorrect_letter data tuple along with game data

{:guess_word, word} - If correct, returns the :correct_word data tuple along with game If incorrect, returns the :incorrect_word data tuple with game data

info(g)

Specs

info(t) :: Keyword.t

Returns Game information

load(id_key, secret, max_wrong)

Specs

load(id, String.t | [String.t], pos_integer) :: t

Loads a new Game state given new secret(s)

mystery_letter()

Specs

mystery_letter :: String.t

status(game)

Specs

status(t) :: tuple

Returns current Game status text