Hangman v0.9.1 Game.Server

Module implements Hangman Game server using GenServer. Wraps Game abstraction as server state. Runs each game or set of games sequentially

Interacts with client player through public interface and maintains game state

Types

id :: String.t

Functions

guess(game_pid, arg)

Specs

guess(pid, Guess.t) :: tuple

Issues guess letter or word request, returns guess result

Internally, runs guess against game secret. Updates Hangman pattern, status, and other game recordkeeping structures.

Guesses follow two types

{:guess_letter, letter} - If correct, returns the :correct_letter data tuple along with game info otherwise, returns the :incorrect_letter data tuple along with game info

{:guess_word, word} - If correct, returns the :correct_word data tuple along with game info If incorrect, returns the :incorrect_word data tuple with game info

load(game_pid, id_key, secret, max_wrong \\ 5)

Specs

load(pid, id, String.t | [String.t], pos_integer) :: :ok

Loads new game into server process. Used primarily by Game.Pid.Cache

secret_length(game_pid)

Specs

secret_length(pid) :: tuple

Retrieves Game secret length

start_link(player_name, secret, max_wrong \\ 5)

Specs

start_link(String.t, String.t | [String.t], pos_integer) :: {:ok, pid}

Start public interface method with secret(s)

status(game_pid)

Specs

status(pid) :: tuple

Retrieves Game status data

stop(game_pid)

Specs

stop(pid) :: tuple

Issues request to stop GenServer

whereis(id_key)

Specs

whereis(id) :: pid | :atom

Routine returns game server pid from process registry using gproc If not found, returns :undefined