Hangman v0.9.1 Game.Server

Module implements hangman game server using GenServer. Wraps hangman game abstraction as server state. Runs each game or set of games sequentially

Interacts with player client through public interface and maintains game state

Types

id :: String.t

Functions

guess(game_pid, arg)

Specs

guess(pid, Guess.t) :: tuple

Issues guess letter or word request, returns guess result reply

Internally, runs guess data against game secret. Updates hangman pattern, status, and other game recordkeeping structures.

Guesses follow two types

{:guess_letter, letter} - If correct, returns the :correct_letter data tuple along with game otherwise, returns the :incorrect_letter data tuple along with game

{:guess_word, word} - If correct, returns the :correct_word data tuple along with game If incorrect, returns the :incorrect_word data tuple with game

load(game_pid, id_key, secret, max_wrong \\ 5)

Specs

load(pid, id, String.t | [String.t], pos_integer) :: :ok

Loads new game state into server process. Used primarily by game pid cache server

secret_length(game_pid)

Specs

secret_length(pid) :: tuple

Retrieves game secret length number

start_link(player_name, secret, max_wrong \\ 5)

Specs

start_link(String.t, String.t | [String.t], pos_integer) :: {:ok, pid}

Start public interface method with secret(s)

status(game_pid)

Specs

status(pid) :: tuple

Retrieves game server status data

stop(game_pid)

Specs

stop(pid) :: tuple

Issues request to stop GenServer

whereis(id_key)

Specs

whereis(id) :: pid | :atom

Routine returns game server pid from process registry using gproc If not found, returns :undefined