Hangman v0.9.5 Hangman.Game

Module to create and manage Hangman game abstractions.

Naturally serves as the symbolic core of the game application.

The game embodies the structure of play in this context of letter and word guessing. Therefore it maintains the secret pattern state along with the guess state, delineated by correct and incorrect, letter and word.

Further, a game wouldn't amount to much without a quantifiable outcome thus we maintain a score state. If a game is lost, the score is automatically 25, as deemed by the threshhold of the particular max incorect guesses.

The game rules are simply to stay within the max incorrect guesses and choose letters consistent with the game. No timing or other such rules are imposed.

If the number of incorrect letter guesses exceeds the threshhold, the game is lost. The lower the score and hence the fewer rounds to guess the word, the more desirable aim.

The Game handles a single Hangman game or multiple Hangman games and runs each game sequentially. Therefore only one game is in play at any one time.

Primary game functions are load/3, guess/2, status/1.

Types

code ::
  :reset |
  :start |
  :guessing |
  :won |
  :lost |
  :finished |
  :abort

Game status code values

feedback :: %{id: String.t, code: code}

returned Game feedback data

id :: String.t

Game id is String.t (currently using player's name as unique key)

t

Functions

abort(game)

empty?(game)

Specs

empty?(t) :: boolean

Returns boolean whether Game is empty

equal?(game1, game2)

Specs

equal?(t, t) :: boolean

Returns boolean whether Game states are identical

guess(game, arg)

Specs

guess(t, guess :: Hangman.Guess.t) :: {t, feedback}

Runs guess against Game secret. Updates Hangman pattern, status, and other game recordkeeping structures.

Guesses follow two types

{:guess_letter, letter} - If correct, returns the :correct_letter data tuple along with game data otherwise, returns the :incorrect_letter data tuple along with game data

{:guess_word, word} - If correct, returns the :correct_word data tuple along with game If incorrect, returns the :incorrect_word data tuple with game data

info(g)

Specs

info(t) :: Keyword.t

Returns Game information

new(id_key, secret, max_wrong)

Specs

new(id, String.t | [String.t], pos_integer) :: t

Creates a new Game state given new secret(s)

next(game)

Specs

next(t) :: {t, term}

secret_length(game)

Specs

secret_length(t) :: integer

status(game)

Specs

status(t) :: {id, code, String.t}

Returns current Game status data and updates status code