Hangman v0.9.1 Hangman.Game.Server behaviour

Module implements hangman game using GenServer process. Wraps hangman game abstraction as server state

Runs each game or set of games sequentially

Interacts with player client through public interface and maintains game state

Types

id :: String.t

Functions

format_status(reason, list)

Specs

format_status(term, [...]) :: [...]

Used for debugging purposes, returns game data of server process

guess(game_pid, arg)

Specs

guess(pid, Guess.t) :: tuple

guess(pid, Guess.t) :: tuple

Issues guess word call to GenServer, returns guess result

load(game_pid, id_key, secret, max_wrong \\ 5)

Specs

load(pid, id, String.t | [String.t], pos_integer) :: :ok

Loads new game state into server process. Used primarily by game pid cache server

secret_length(game_pid)

Specs

secret_length(pid) :: tuple

Retrieves game secret length number

start_link(player_name, secret, max_wrong \\ 5)

Specs

start_link(String.t, String.t | [String.t], pos_integer) :: {:ok, pid}

Start public interface method with secret

status(game_pid)

Specs

status(pid) :: tuple

Retrieves game server status data

stop(game_pid)

Specs

stop(pid) :: tuple

Issues request to stop GenServer

whereis(id_key)

Specs

whereis(id) :: pid | :atom

Routine returns game server pid from process registry using gproc If not found, returns :undefined

Callbacks

handle_call(atom, tuple, arg2)

Specs

handle_call(:atom, tuple, Hangman.Game.t) :: tuple

handle_call(:atom, tuple, Hangman.Game.t) :: tuple

handle_call(:atom, tuple, Hangman.Game.t) :: tuple

handle_call(Guess.t, tuple, Hangman.Game.t) :: tuple

handle_call(Guess.t, tuple, Hangman.Game.t) :: tuple

Stops the server is a normal graceful way

handle_cast(tuple, arg1)

Specs

handle_cast(tuple, Hangman.Game.t) :: tuple

handle_cast(tuple, Hangman.Game.t) :: tuple

Loads a set of games

init(arg0)

Specs

init(Hangman.Game.t) :: tuple

GenServer callback to initalize server process

terminate(term, term)

Specs

terminate(term, term) :: :ok

Terminates the hangman game server No special cleanup other than refreshing the state