Hangman v0.9.5 Hangman.Letter.Strategy

Module provides access to a set of functions handling letter guessing strategy.

(From Wikipedia)

In the English language, the twelve most commonly occurring letters are, in descending order: e-t-a-o-i-n-s-h-r-d-l-u. This and other letter-frequency lists are used by the guessing player to increase the odds when it is their turn to guess. On the other hand, the same lists can be used by the puzzle setter to stump their opponent by choosing a word which deliberately avoids common letters (e.g. rhythm or zephyr) or one that contains rare letters (e.g. jazz).

Another common strategy is to guess vowels first, as English only has five vowels (a, e, i, o, and u, while y may sometimes, but rarely, be used as a vowel) and almost every word has at least one.

According to a 2010 study conducted by Jon McLoone for Wolfram Research, the most difficult words to guess include jazz, buzz, hajj, faff, fizz, fuzz and variations of these.

NOTE: For the implementation, letter-frequency lists are primarily used. A vowel strategy is not explicitly used, and perhaps is utilized implicitly through letter-frequency lists. No effort at this point is made to stump the player.

For the robot player type, the strategy retrieves the best letter considering English letter frequencies and letter tally counts.

For the human player type, the strategy retries a set of top letter choices to be presented to human to manually choose.

Types

result :: {t, Hangman.Guess.t}

t

Functions

choices(strategy, round)

Specs

choices(t, Hangman.Round.t) :: Hangman.Guess.options

Retrieves choices info augmenting it with round info

guess(strategy)

Specs

guess(t) :: Hangman.Guess.t

Returns best letter guess as previously computed

guess(strategy, atom)

Specs

guess(t, atom) :: Hangman.Guess.t

guess(strategy, atom, arg)

Specs

guess(t, atom, tuple) :: Hangman.Guess.t

guessed(strategy)

Specs

guessed(t) :: list

Returns a list of guessed letters up to this point

info(s)

Specs

info(t) :: Keyword.t

Returns Strategy information

new()

Specs

new :: t

Returns a new Letter.Strategy

process(strategy, pass)

Specs

process(t, Hangman.Pass.t) :: t | no_return

Process best letter guess as deemed by the auto self-selecting heuristic

process(strategy, atom, pass)

Specs

process(t, atom, Hangman.Pass.t) :: t | no_return

process(strategy, atom, pass, n \\ 5)

Specs

process(t, atom, Hangman.Pass.t, pos_integer) ::
  Hangman.Guess.option |
  no_return

Processes pass data and stores guess choices data. Denotes 'strategic' letter pick with an asterisk.

update(strategy, arg)

Specs

update(t, Hangman.Guess.t) :: t

Updates strategy with guess data