Hangman v0.9.5 Hangman.Player.FSM

Module implements a non-process player fsm which handles managing the state of types implemented through Player and the Player Action protocol.

FSM provides a state machine wrapper over the Player

The FSM is not coupled at all to the specific player type but the generic Player which relies on the Action Protocol, which provides for succinct code along with the already succinct design of the Fsm module code.

Works for all supported player types

States are initial, begin, setup, action, transit, exit

The event proceed transitions between states, when we are not issuing a guess or initialize event.

Here are the state transition flows:

initial -> begin

begin -> setup | exit

setup -> action

action -> transit | setup

transit -> begin | exit

exit -> exit

Basically upon leaving the initial state, we transition to begin. From there we make the determination of whether we should proceed on to setup the guess state or terminate early and exit.

If the game was recently just aborted and we are done with playing any more games -> we exit.

Once we are in the setup state it is obvious that our next step is to the action state. Here we can try out our new guess (either selected or auto-generated)

From action state we either circle back to setup state to generate the new word set state and overall guess state and possibly to collect the next user guess. Else, we have either won or lost the game and can confidently move to the transit state.

The transit state indicates that we are in transition having a single game over. Either we proceed to start a new game and head to begin or we've already finished all games and happily head to the exit state.

Ultimately the specific Flow or CLI Handler, when in exit state terminates the FSM loop

Functions

data(f_s_m)

guess(fsm, arg0)

initialize(fsm, arg0)

new()

proceed(fsm)

state(f_s_m)

transition(fsm, , )