Hangman v0.9.1 Player

Module handles player abstraction and defines player type

Implements functionality to start a player, choose letters, guess letters and words

Heavily relies upon Player.Round and Guess.Action functionality

Types

kind :: :human | :robot

result :: {t, Player.Round.result}

t :: %Player{event_server_pid: term, game_no: term, game_server_pid: term, games_summary: term, mystery_letter: term, name: term, round: term, round_no: term, strategy: term, type: term}

Functions

choices(p, atom, options \\ nil)

Specs

choices(t, Guess.directive, :atom) :: {t, Guess.option}

HUMAN Routine for human player type retrieves letter choices Setups new round, retrieves and returns letter choices

delete(p)

Specs

delete(t) :: t

Method returns empty player state

game_lost?(p)

Specs

game_lost?(t) :: boolean

Returns true or false whether the game has been lost

game_won?(p)

Specs

game_won?(t) :: boolean

Returns true or false whether the game has been won

games_over?(p)

Specs

games_over?(t) :: boolean

Returns true or false whether all games are over

games_summary(args)

Specs

games_summary(Keyword.t) :: String.t

Returns game summary as a string

guess(p)

Specs

guess(t) :: result

guess(t) :: result

ROBOT Routine for robot player type guess Setups new round, performs guess, returns round status

guess(p, arg2)

Specs

guess(t, Guess.t) :: result

guess(t, Guess.directive) :: result

HUMAN Routine for human player type which perform letter guesses

Doesn’t setup round since it was setup during choose letters stage. Issues action to guess letter and returns round status

human()

info(p)

Specs

info(t) :: Keyword.t

last_word?(p)

Specs

last_word?(t) :: boolean

Returns true or false whether we’ve arrived at the last word in possible words set

new(name, type, game_pid, event_pid)

Specs

new(String.t, :atom, pid, pid) :: t

Returns new player, validates player type

robot()

start(p)

Specs

start(t) :: result

Routine starts a new player abstraction.

Notifies player specific event server

Setups game round.

If player type is robot, makes first guess, and returns round status If player type is human retrieves letter choices to display

status(p, atom)

Specs

status(t, :atom) :: Player.Round.result

status(t, :atom) :: Player.Round.result

Returns games over status