Hangman v0.9.1 Player.Game

Module handles relationship between player, game server and player event server

Supports both synchronous human and synchronous robot types

Loads up player specific game components : dynamic game server, and event server given player

Manages specific player fsm behaviour (human or robot).

Wraps fsm game play into an enumerable for easy running.

Functions

rounds_handler(arg, atom)

Specs

rounds_handler(tuple, Player.kind) :: Enumerable.t

Permits robot or human round playing! Wraps the robot or human player game playing in a stream. Stream resource returns an enumerable.

Terminates player events server and fsm upon finish

run(name, type, secrets, log, display)

Specs

run(String.t, Player.kind, [String.t], boolean, boolean) :: :ok

Function run connects all the player specific components together and runs the player game

setup(name, secrets, log, display)

Specs

setup(String.t, [String.t], boolean, boolean) :: tuple

Function setup loads the player specific game components. Setup the game server and per player event server.

start_player(name, type, game_pid, notify_pid)

Specs

start_player(String.t, Player.kind, pid, pid) :: Supervisor.on_start_child

Start dynamic player child worker