Hangman v0.9.1 Round

Module to implement Hangman game round abstraction.

Works in conjuction with Strategy and Game.Server to orchestrate actual round game play.

Basic Round functionality includes setup/2, guess/2, update/3, status/1

Types

result :: {status_code :: :atom, status_txt :: String.t}

t :: %Round{final_result: term, guess: term, pattern: term, result_code: term, seq_no: term, status_code: term, status_text: term}

Functions

augment_choices(player, arg)

Specs

augment_choices(Player.t, Guess.option) :: Guess.option

Interjects round specific parameters into choices text.

context(player)

Specs

context(Player.t) :: Guess.context | no_return

Returns round context based on results of last guess

guess(p, arg)

Specs

guess(Player.t, Guess.t) :: t

Issues a client guess (either letter or word) against Game.Server. Notifies player events of guess results.

Returns received round data

setup(player)

Setups game play round

For game start stage, retrieves secret length from game server uses secret length to filter possible Hangman words from Pass.Cache server

On subsequent rounds, generates a reduce key based on the result of the last guess to filter possible Hangman words from Pass.Cache server

setup(player, context)

Specs

setup(Player.t, none | :atom | Guess.context) :: Player.t

status(player)

Specs

status(Player.t) :: result

Returns round status tuple

update(player, round)

update(player, round, arg)

Specs

update(Player.t, Round.t, none | Guess.t) :: Player.t

Updates Player abstraction with Round results. If games are over, updates games summary and notifies Player.Events.

Under human guessing, round update will update the strategy with the guess particulars. If robot guessing, the strategy will also be updated.