Hangman v0.9.1 Round

Module provides access to Hangman guessing round functions.

Works in conjuction with Strategy and Game.Server to orchestrate actual round game play.

When playing a Hangman game, we first setup our round, which involves obtaining the secret word length from the game server. Next, we take steps to reduce the possible Hangman words set to narrow our word choices. From there on we choose the best letter to guess given the knowledge we have of our words data set. If we are a :human', we are given letter choices to pick, and if we are a ':robot', we trust our friend [Strategy](Strategy.html). After a guess is made either by:humanor:robot` we update our round recordkeeping structures with the guess results and proceed for the next round -- to do it all over again.

Basic Round functionality includes setup/2, guess/2, update/3, status/1.

Types

code ::
  :correct_letter |
  :incorrect_letter |
  :incorrect_word

result :: {code, status :: String.t}

t :: %Round{final_result: term, guess: term, pattern: term, result_code: term, seq_no: term, status_code: term, status_text: term}

Functions

augment_choices(player, arg)

Specs

augment_choices(Player.t, Guess.option) :: Guess.option

Interjects round specific parameters into choices text.

context(player)

Specs

context(Player.t) :: Guess.context | no_return

Returns round context based on results of last guess

guess(p, arg)

Specs

guess(Player.t, Guess.t) :: t

Issues a client guess (either letter or word) against Game.Server. Notifies player events of guess results.

Returns received round data

setup(player)

Setups game play round

For game start stage, retrieves secret length from game server uses secret length to filter possible Hangman words from Pass.Cache server

On subsequent rounds, generates a reduce key based on the result of the last guess to filter possible Hangman words from Pass.Cache server

setup(player, context)

Specs

setup(Player.t, none | :atom | Guess.context) :: Player.t

status(player)

Specs

status(Player.t) :: result

Returns round status tuple

update(player, round)

update(player, round, arg)

Specs

update(Player.t, Round.t, none | Guess.t) :: Player.t

Updates Player abstraction with Round results. If games are over, updates games summary and notifies Player.Events.

Under human guessing, round update will update the strategy with the guess particulars. If robot guessing, the strategy will also be updated.