Hangman v0.9.1 Strategy

Handles letter guessing strategy for player.

For robot player type, retrieves best letter considering english letter frequencies and letter tally counts.

For human player type, retries a set of top letter choices to be presented to human to manually choose.

Types

result :: {t, Guess.t}

t :: %Strategy{guess: term, guessed_letters: term, pass: term, prior_guess: term}

Functions

choose_letters(strategy, n \\ 5)

Specs

choose_letters(t, pos_integer) :: Guess.option

Retrieves letter choices text. Denotes ‘strategic’ letter pick with an asterisk.

get_guessed(strategy)

Specs

get_guessed(t) :: list

Returns a list of guessed letters up to this point

info(s)

Specs

info(t) :: Keyword.t

Returns strategy information

last_word(strategy)

Specs

last_word(t) :: Guess.t

If we have reached the last possible hangman word, return it so we can guess it

letter_in_most_common(strategy, letter, n \\ 5)

Specs

letter_in_most_common(t, pos_integer, String.t) :: Guess.t

Validates letter is within the top strategy letter choices. If not, picks the top letter as deemed by heuristics.

make_guess(strategy)

Specs

make_guess(t) :: Guess.t

Returns best letter guess as deemed by strategy heuristics

most_common_letter(strategy, n)

Specs

most_common_letter(t, pos_integer) :: list

Returns most common n letters

most_common_letter_and_counts(strategy, n)

Specs

most_common_letter_and_counts(t, pos_integer) :: Keyword.t

Returns most common n letters along with their counts

new()

Specs

new :: t

Returns a new strategy

retrieve_best_letter(strategy)

Specs

retrieve_best_letter(t) :: String.t

Method implements the most common letter retrieval strategy with a twist. Gets the first letter with the highest frequency for when the current possible hangman word set space is > “small”. The twist is when we combine the english language letter relative frequency. For the cases where the word set is less than small, takes the letter whose frequencies are less than or equal to half the possible hagman word pass size.

E.g.for size 10, the letter counts would need to be 5 or lower to be chosen. Doesn’t handle tie between letters.

update(strategy, arg)

Specs

update(t, Guess.t) :: t

Updates strategy with guess particulars

update(strategy, pass, atom)

Specs

update(t, Pass.t, Player.kind) :: t

Updates strategy with pass data.

For robot player type, when we updated the pass data also prepare the guess.